### 2021-06-12 #3360 @rusu24edward
Potential relevant reviewers, as per whedon's instructions above: satpreetsingh, seba-1511, abhiramm7, bozutemiz, mhetrerajat, giodegas, ionlights.

I'm not sure why the pdf did not build because it builds fine in my GitHub Action. I also don't understand the missing DOI error...

Thanks!
