### 2021-06-14 87eeba1
Walls block agent movement
