### 2021-06-14 ce2e2f1
prototype dynamic order manager with small example
