### 2021-06-14 4df3731
adding some logic for ray-obs masking
