+++ abmarl/sim/simplified_grid.py
@@ -91,7 +91,181 @@ class GridSim(AgentBasedSimulation):
         """
         Return the agent's observation.
         """
-        pass
+        agent = self.agents[agent_id]
+        # Generate a completely empty grid
+        view = np.empty((agent.view_range * 2 + 1, agent.view_range * 2 + 1), dtype=object)
+        # TODO: distinguish between empty out of bounds and empty in bounds.
+        
+        # Copy the section of the grid around the agent's location
+        (r, c) = agent.position
+        r_lower = max([0, r - agent.view_range])
+        r_upper = min([self.rows - 1, r + agent.view_range]) + 1
+        c_lower = max([0, c - agent.view_range])
+        c_upper = min([self.cols - 1, c+agent.view_range]) + 1
+        signal[
+            (r_lower+agent.view_range-r):(r_upper+agent.view_range-r),
+            (c_lower+agent.view_range-c):(c_upper+agent.view_range-c)
+        ] = self.grid[r_lower:r_upper, c_lower:c_upper]
+
+        factor = np.sqrt(0.5)
+    
+        # Generate a mask around the agent's location. The mask will filter out
+        # elements of the grid by view-blocking agents.
+        for other in self.agents.values():
+            if isinstance(other, WallAgent):
+                # get relative position between agent and wall
+                rel_pos = other.position - agent.position
+                if rel_pos[0] == 0 and rel_pos[1] > 0: # Wall is right of agent
+                    upper_point = (rel_pos[1] - factor, rel_pos[0] - factor)
+                    lower_point = (rel_pos[1] - factor, rel_pos[0] + factor)
+                elif rel_pos[0] < 0 and rel_pos[1] > 0: # Wall is above-right of agent
+                    upper_point = (rel_pos[1] - factor, rel_pos[0] - factor)
+                    lower_point = (rel_pos[1] + factor, rel_pos[0] + factor)
+                elif rel_pos[0] < 0  and rel_pos[1] == 0 : # Wall is above the agent
+                    upper_point = (rel_pos[1] - factor, rel_pos[0] + factor)
+                    lower_point = (rel_pos[1] + factor, rel_pos[0] + factor)
+                elif rel_pos[0] < 0 and rel_pos[1] < 0: # Wall is above-left of agent
+                    upper_point = (rel_pos[1] - factor, rel_pos[1] + factor)
+                    lower_point = (rel_pos[1] + factor, rel_pos[1] - factor)
+                elif rel_pos[0] == 0 and rel_pos[1] < 0: # Wall is left of agent
+                    upper_point = (rel_pos[1] + factor, rel_pos[0] - factor)
+                    lower_point = (rel_pos[1] + factor, rel_pos[0] + factor)
+                elif rel_pos[0] > 0 and rel_pos[1] < 0: # Wall is below-left of agent
+                    upper_point = (rel_pos[1] - factor, rel_pos[0] - factor)
+                    lower_point = (rel_pos[1] + factor, rel_pos[0] + factor)
+                elif rel_pos[0] > 0 and rel_pos[1] == 0: # Wall is below the agent
+                    upper_point = (rel_pos[1] - factor, rel_pos[0] - factor)
+                    lower_point = (rel_pos[1] + factor, rel_pos[0] - factor)
+                elif rel_pos[0] > 0 and rel_pos[1] > 0: # Wall is below-right of agent
+                    upper_point = (rel_pos[1] + factor, rel_pos[0] - factor)
+                    lower_point = (rel_pos[1] - factor, rel_pos[0] + factor)
+
+                # generate two lines from agent center to wall corners
+                # all points/cells between those two lines and behind the wall are masked off.
+
+
+
+
[…]
