### 2021-06-14 35c16df
More logic for grid observation masking
