### 2021-06-15 36f1957
Observation masking checks
