+++ abmarl/sim/simplified_grid.py
@@ -107,145 +107,78 @@ class GridSim(AgentBasedSimulation):
             (c_lower+agent.view_range-c):(c_upper+agent.view_range-c)
         ] = self.grid[r_lower:r_upper, c_lower:c_upper]
 
-        local_visible = np.ones((agent.view_range * 2 + 1, agent.view_range * 2 + 1))
-
-        factor = np.sqrt(0.5)
-    
-        # Generate a mask around the agent's location. The mask will filter out
-        # elements of the grid by view-blocking agents.
-        for other in self.agents.values():
-            if isinstance(other, WallAgent):
-                # get relative position between agent and wall
-                rel_pos = other.position - agent.position
-                if rel_pos[0] == 0 and rel_pos[1] > 0: # Wall is right of agent
-                    upper_point = (rel_pos[1] - factor, rel_pos[0] - factor)
-                    lower_point = (rel_pos[1] - factor, rel_pos[0] + factor)
-                elif rel_pos[0] < 0 and rel_pos[1] > 0: # Wall is above-right of agent
-                    upper_point = (rel_pos[1] - factor, rel_pos[0] - factor)
-                    lower_point = (rel_pos[1] + factor, rel_pos[0] + factor)
-                elif rel_pos[0] < 0  and rel_pos[1] == 0 : # Wall is above the agent
-                    upper_point = (rel_pos[1] - factor, rel_pos[0] + factor)
-                    lower_point = (rel_pos[1] + factor, rel_pos[0] + factor)
-                elif rel_pos[0] < 0 and rel_pos[1] < 0: # Wall is above-left of agent
-                    upper_point = (rel_pos[1] - factor, rel_pos[1] + factor)
-                    lower_point = (rel_pos[1] + factor, rel_pos[1] - factor)
-                elif rel_pos[0] == 0 and rel_pos[1] < 0: # Wall is left of agent
-                    upper_point = (rel_pos[1] + factor, rel_pos[0] - factor)
-                    lower_point = (rel_pos[1] + factor, rel_pos[0] + factor)
-                elif rel_pos[0] > 0 and rel_pos[1] < 0: # Wall is below-left of agent
-                    upper_point = (rel_pos[1] - factor, rel_pos[0] - factor)
-                    lower_point = (rel_pos[1] + factor, rel_pos[0] + factor)
-                elif rel_pos[0] > 0 and rel_pos[1] == 0: # Wall is below the agent
-                    upper_point = (rel_pos[1] - factor, rel_pos[0] - factor)
-                    lower_point = (rel_pos[1] + factor, rel_pos[0] - factor)
-                elif rel_pos[0] > 0 and rel_pos[1] > 0: # Wall is below-right of agent
-                    upper_point = (rel_pos[1] + factor, rel_pos[0] - factor)
-                    lower_point = (rel_pos[1] - factor, rel_pos[0] + factor)
+        mask = np.ones((2 * agent.view + 1, 2 * agent.view + 1))
 
         for other in self.agents.values():
             if isinstance(other, WallAgent):
                 x_diff = other.position[1] - agent.position[1]
                 y_diff = agent.position[0] - other.position[0]
-                if 0 < x_diff < agent.view_range and 0 < y_diff < agent.view_range:
-                    upper_line = lambda x: (y_diff + factor) / (x_diff - factor) * x
-                    lower_line = lambda x: (y_diff - factor) / (x_diff + factor) * x
-                    for x in range(x_diff, agent.view_range):
-                        for y in range(y_diff, agent.view_range):
-                            if lower_line(x) < y < upper_line(x):
-                                mask[x, y] = 0
+                if -agent.view <= x_diff <= agent.view and \
[…]
