### 2021-06-15 7075e0c
Right of case converted to row column logic
