### 2021-06-15 85db668
Updated below-right logic
