### 2021-06-15 783f146
Fixed bug with wall blocking another view-blocking wall
