+++ abmarl/sim/filter_obs.py
@@ -5,23 +5,19 @@ view = 8
 grid = np.ones((2 * view + 1, 2 * view + 1))
 grid[view, view] = 0 # agent's location
 
-x_diffs = [4, -3, 3, -2, -5]
-y_diffs = [-3, 0, 2, 3, 0]
+r_diffs = [3, 0, -2, -3, 0, -3, -4]
+c_diffs = [4, -3, 3, -2, -5, 3, 3]
 
-for i in range(1):
-    # x_diff = x_diffs[i]
-    # y_diff = y_diffs[i]
-
-    r_diff = -1
-    c_diff = 3
-
-    if grid[r_diff + view, c_diff + view] != 1: continue # I cannot see the wall because its behind another wall
+for i in range(len(r_diffs)):
+    r_diff = r_diffs[i]
+    c_diff = c_diffs[i]
 
     if c_diff > 0 and r_diff == 0: # Wall is to the right of agent
         upper = lambda t: (r_diff + 0.5) / (c_diff - 0.5) * t
         lower = lambda t: (r_diff - 0.5) / (c_diff - 0.5) * t
         for c in range(c_diff, view+1):
             for r in range(-view, view+1):
+                if c == c_diff and r == r_diff: continue # don't mask the wall
                 if lower(c) < r < upper(c):
                     grid[r + view, c + view] = 0
 
@@ -31,6 +27,7 @@ for i in range(1):
         lower = lambda t: (r_diff - 0.5) / (c_diff + 0.5) * t
         for c in range(c_diff, view+1):
             for r in range(r_diff, view+1):
+                if c == c_diff and r == r_diff: continue # Don't mask the wall
                 if lower(c) < r < upper(c):
                     grid[r + view, c + view] = 0
 
@@ -40,6 +37,7 @@ for i in range(1):
         right = lambda t: (c_diff + 0.5) / (r_diff - 0.5) * t
         for c in range(-view, view+1):
             for r in range(r_diff, view+1):
+                if c == c_diff and r == r_diff: continue # don't mask the wall
                 if left(r) < c < right(r):
                     grid[r + view, c + view] = 0
 
@@ -49,6 +47,7 @@ for i in range(1):
         lower = lambda t: (r_diff - 0.5) / (c_diff - 0.5) * t
         for c in range(c_diff, -view-1, -1):
             for r in range(r_diff, view+1):
+                if c == c_diff and r == r_diff: continue # don't mask the wall
                 if lower(c) < r < upper(c):
                     grid[r + view, c + view] = 0
 
@@ -58,6 +57,7 @@ for i in range(1):
         lower = lambda t: (r_diff - 0.5) / (c_diff + 0.5) * t
         for c in range(c_diff, -view-1, -1):
             for r in range(-view, view+1):
+                if c == c_diff and r == r_diff: continue # don't mask the wall
                 if lower(c) < r < upper(c):
                     grid[r + view, c + view] = 0
 
@@ -67,6 +67,7 @@ for i in range(1):
         lower = lambda t: (r_diff - 0.5) / (c_diff + 0.5) * t
         for c in range(c_diff, -view - 1, -1):
             for r in range(r_diff, -view - 1, -1):
+                if c == c_diff and r == r_diff: continue # don't mask the wall
                 if lower(c) < r < upper(c):
                     grid[r + view, c + view] = 0
 
@@ -76,6 +77,7 @@ for i in range(1):
         right = lambda t: (c_diff + 0.5) / (r_diff + 0.5) * t
         for c in range(-view, view+1):
             for r in range(r_diff, -view - 1, -1):
+                if c == c_diff and r == r_diff: continue # don't mask the wall
                 if left(r) < c < right(r):
[…]
