### 2021-06-15 c050d83
Moved observation filter logic to simplified grid
