### 2021-06-15 076f204
Convolving the grid with the mask
