### 2021-06-15 d9b566a
Fixed mask bug and added object encoding for observation
