+++ abmarl/sim/simplified_grid.py
@@ -1,9 +1,11 @@
 
+from abc import abstractproperty
+
 from gym.spaces import Box
 from matplotlib import pyplot as plt
 import numpy as np
 
-from abmarl.sim import PrincipleAgent, ActingAgent, AgentBasedSimulation
+from abmarl.sim import PrincipleAgent, Agent, ActingAgent, AgentBasedSimulation
 from abmarl.tools.matplotlib_utils import mscatter
 
 class GridAgent(PrincipleAgent):
@@ -14,15 +16,27 @@ class GridAgent(PrincipleAgent):
     @position.setter
     def position(self, value):
         self._position = value
+    
+    @abstractproperty
+    def encode(self):
+        pass
 
-class WallAgent(GridAgent): pass
+class WallAgent(GridAgent):
+    @property
+    def encode(self):
+        return 2
 
-class ExploringAgent(GridAgent, ActingAgent):
+class ExploringAgent(GridAgent, Agent):
     def __init__(self, view_range=None, move_range=None, **kwargs):
         super().__init__(**kwargs)
         self.view_range = view_range
         self.move_range = move_range
         self.action_space = Box(-move_range, move_range, (2,), np.int)
+        self.observation_space = Box(-np.inf, np.inf, (view_range, view_range), np.int)
+    
+    @property
+    def encode(self):
+        return 1
 
 
 
@@ -110,87 +124,90 @@ class GridSim(AgentBasedSimulation):
         # Generate an observation mask. The agent's observation can be blocked
         # by walls, which hide the cells "behind" them. In the mask, 1 means that
         # this square is visibile, 0 means that it is invisible.
-        mask = np.ones((2 * agent.view + 1, 2 * agent.view + 1))
+        mask = np.ones((2 * agent.view_range + 1, 2 * agent.view_range + 1))
         for other in self.agents.values():
             if isinstance(other, WallAgent):
                 r_diff, c_diff = other.position - agent.position
-                if -agent.view <= r_diff <= agent.view and \
-                        agent.view <= c_diff <= agent.view:
+                if -agent.view_range <= r_diff <= agent.view_range and \
+                        -agent.view_range <= c_diff <= agent.view_range:
                     if c_diff > 0 and r_diff == 0: # Wall is to the right of agent
                         upper = lambda t: (r_diff + 0.5) / (c_diff - 0.5) * t
                         lower = lambda t: (r_diff - 0.5) / (c_diff - 0.5) * t
-                        for c in range(c_diff, agent.view+1):
-                            for r in range(-agent.view, agent.view+1):
+                        for c in range(c_diff, agent.view_range+1):
+                            for r in range(-agent.view_range, agent.view_range+1):
                                 if c == c_diff and r == r_diff: continue # don't mask the wall
                                 if lower(c) < r < upper(c):
-                                    mask[r + agent.view, c + agent.view] = 0
+                                    mask[r + agent.view_range, c + agent.view_range] = 0
                     elif c_diff > 0 and r_diff > 0: # Wall is below-right of agent
                         upper = lambda t: (r_diff + 0.5) / (c_diff - 0.5) * t
                         lower = lambda t: (r_diff - 0.5) / (c_diff + 0.5) * t
-                        for c in range(c_diff, agent.view+1):
-                            for r in range(r_diff, agent.view+1):
+                        for c in range(c_diff, agent.view_range+1):
+                            for r in range(r_diff, agent.view_range+1):
                                 if c == c_diff and r == r_diff: continue # Don't mask the wall
                                 if lower(c) < r < upper(c):
-                                    mask[r + agent.view, c + agent.view] = 0
+                                    mask[r + agent.view_range, c + agent.view_range] = 0
                     elif c_diff == 0 and r_diff > 0: # Wall is below the agent
[…]
