### 2021-06-15 dc1dba3
Customizable encoding per object
