### 2021-06-15 e84d322
Restricted certain encoding values
