### 2021-06-15 2e27747
Grid agents have initial positions and GridSim still unique positions
