### 2021-06-15 84479f8
Randomly assign agents position only if they don't have initial position
