### 2021-06-15 75cc7ad
Build the grid from a startup configuration file
