+++ .github/workflows/build-and-test.yml
@@ -6,7 +6,7 @@ on:
   push:
     branches: [ main ]
   pull_request:
-    branches: [ main ]
+    branches: [ main, abmarl-epic-simplified-grid-world ]
 
 jobs:
   build-and-test:
