### 2021-06-15 1878e46
Split concrete agents into abstract feature classes
