### 2021-06-15 c056502
Added render shape property
