+++ abmarl/sim/build_grid_example.py
@@ -1,8 +1,19 @@
 
 from matplotlib import pyplot as plt
 
-from abmarl.sim.simplified_grid import WallAgent, ExploringAgent, build_grid_sim
+from abmarl.sim.simplified_grid import GridAgent, MovingAgent, GridObservingAgent, build_grid_sim
 
+class WallAgent(GridAgent):
+    """
+    Custom WallAgent with default encoding as 1.
+    """
+    def __init__(self, encode=1, **kwargs):
+        super().__init__(**{'encode': encode, **kwargs})
+
+class ExploringAgent(MovingAgent, GridObservingAgent):
+    def __init__(self, encode=2, **kwargs):
+        super().__init__(**{'encode': encode, **kwargs})
+        
 object_registry = {
     'A': lambda n: ExploringAgent(id=f'explorer{n}', view_range=3, move_range=1),
     'W': lambda n: WallAgent(id=f'wall{n}')
