### 2021-06-15 86ec86d
Cleaned up encoding and render shape
