### 2021-06-16 93d7144
Moved agents to own file, making simplified grid more like an example.
