### 2021-06-16 4ad104a
More weight to GridlWorldBaseComponent
