### 2021-06-16 2af79f5
grid state handles the details of modiftying agent positions
