+++ abmarl/sim/grid_world/state.py
@@ -49,3 +49,17 @@ class GridWorldState(GridWorldBaseComponent):
                 c = cs[ndx]
                 agent.position = np.array([r, c])
                 self.grid[r, c] = agent
+
+    def modify_position(self, agent, new_position, **kwargs):
+        """
+        Attempt to assign a new position to an agent.
+
+        The new position must be in bounds and must not be occupied by another
+        agent.
+        """
+        if 0 <= new_position[0] < self.rows and \
+                0 <= new_position[1] < self.cols and \
+                self.grid[new_position[0], new_position[1]] is None:
+            self.grid[agent.position[0], agent.position[1]] = None
+            agent.position = new_position
+            self.grid[agent.position[0], agent.position[1]] = agent
