### 2021-06-16 70243d3
Replaced grid world example with state and actor components
