### 2021-06-16 245b792
Added observer. Agents now have a view_blocking property
