+++ abmarl/sim/grid_world/observer.py
@@ -1,0 +1,154 @@
+
+from abc import ABC, abstractmethod
+
+import numpy as np
+
+from abmarl.sim.grid_world import GridWorldBaseComponent, GridObservingAgent
+from abmarl.sim.grid_world.state import GridWorldState
+
+class ObserverBaseComponent(GridWorldBaseComponent, ABC):
+    @abstractmethod
+    def get_obs(self, agent, **kwargs): pass
+
+    @property
+    @abstractmethod
+    def key(self): pass
+
+
+class GridObserver(ObserverBaseComponent):
+    def __init__(self, grid_state=None, **kwargs):
+        super().__init__(**kwargs)
+        self.grid_state = grid_state
+    
+    def get_obs(self, agent, **kwargs):
+        """
+        The agent observes a sub-grid centered on its position.
+
+        The observation contains other agents, empty spaces, out of bounds, and
+        masked cells, which are blocked from view by walls.
+
+        Returns:
+            The observation as a dictionary keyed on this class's key.
+        """
+        if not isinstance(agent, GridObservingAgent):
+            return {}
+        
+        # Generate a completely empty grid
+        local_grid = np.empty((agent.view_range * 2 + 1, agent.view_range * 2 + 1), dtype=object)
+        local_grid.fill(-1)
+
+        # Copy the section of the grid around the agent's location
+        (r, c) = agent.position
+        r_lower = max([0, r - agent.view_range])
+        r_upper = min([self.grid_state.rows - 1, r + agent.view_range]) + 1
+        c_lower = max([0, c - agent.view_range])
+        c_upper = min([self.grid_state.cols - 1, c + agent.view_range]) + 1
+        local_grid[
+            (r_lower+agent.view_range-r):(r_upper+agent.view_range-r),
+            (c_lower+agent.view_range-c):(c_upper+agent.view_range-c)
+        ] = self.grid_state.grid[r_lower:r_upper, c_lower:c_upper]
+
+        # Generate an observation mask. The agent's observation can be blocked
+        # by walls, which hide the cells "behind" them. In the mask, 1 means that
+        # this square is visibile, 0 means that it is invisible.
+        mask = np.ones((2 * agent.view_range + 1, 2 * agent.view_range + 1))
+        for other in self.agents.values():
+            if other.view_blocking:
+                r_diff, c_diff = other.position - agent.position
+                if -agent.view_range <= r_diff <= agent.view_range and \
+                        -agent.view_range <= c_diff <= agent.view_range:
+                    if c_diff > 0 and r_diff == 0: # Wall is to the right of agent
+                        upper = lambda t: (r_diff + 0.5) / (c_diff - 0.5) * t
+                        lower = lambda t: (r_diff - 0.5) / (c_diff - 0.5) * t
+                        for c in range(c_diff, agent.view_range+1):
+                            for r in range(-agent.view_range, agent.view_range+1):
+                                if c == c_diff and r == r_diff: continue # don't mask the wall
+                                if lower(c) < r < upper(c):
+                                    mask[r + agent.view_range, c + agent.view_range] = 0
+                    elif c_diff > 0 and r_diff > 0: # Wall is below-right of agent
+                        upper = lambda t: (r_diff + 0.5) / (c_diff - 0.5) * t
+                        lower = lambda t: (r_diff - 0.5) / (c_diff + 0.5) * t
+                        for c in range(c_diff, agent.view_range+1):
+                            for r in range(r_diff, agent.view_range+1):
+                                if c == c_diff and r == r_diff: continue # Don't mask the wall
+                                if lower(c) < r < upper(c):
+                                    mask[r + agent.view_range, c + agent.view_range] = 0
[…]
