### 2021-06-16 f086bf4
Logic check in MoveActor
