### 2021-06-16 ed7c01b
Cleaned up docstring in Gridworldbasecomponent
