### 2021-06-16 e64405d
Docstrings for agent classes.
