### 2021-06-16 87d3c2e
Docstrings for observer and work out a few more details
