+++ abmarl/sim/gridworld/state.py
@@ -7,7 +7,7 @@ from abmarl.sim.gridworld import GridWorldBaseComponent
 
 class StateBaseComponent(GridWorldBaseComponent, ABC):
     """
-    Abstract State Component base from which all state components wil inherit.
+    Abstract State Component base from which all state components will inherit.
     """
     @abstractmethod
     def reset(self, **kwargs):
