### 2021-06-16 ba42882
Actors and Observers once again assign the relevant spaces to the agent
