### 2021-06-16 4425452
build_sim_grid lives in the gridworld utilties function.
