### 2021-06-16 1c41f45
Updated build grid example
