+++ abmarl/sim/gridworld/__init__.py
@@ -3,7 +3,7 @@ from abc import ABC
 
 import numpy as np
 
-from .agent import GridWorldAgent, GridObservingAgent,  MovingAgent
+from .agent import GridWorldAgent, GridObservingAgent, MovingAgent
 from .util import build_grid_sim
 
 
@@ -15,7 +15,7 @@ class GridWorldBaseComponent(ABC):
     """
     def __init__(self, agents=None, **kwargs):
         self.agents = agents
-    
+
     @property
     def agents(self):
         """
