+++ abmarl/sim/gridworld/build_grid_example.py
@@ -3,7 +3,7 @@ from matplotlib import pyplot as plt
 
 from abmarl.sim.gridworld import build_grid_sim
 from abmarl.sim.gridworld.gridworld_example import WallAgent, ExploringAgent, GridSim
-        
+
 object_registry = {
     'A': lambda n: ExploringAgent(id=f'explorer{n}', view_range=3, move_range=1),
     'W': lambda n: WallAgent(id=f'wall{n}')
