### 2021-06-17 702e31e
Lots of touch ups in the grid world framework
