### 2021-06-17 8f85fbe
Simplified what is in the __init__
