### 2021-06-17 4b2720b
Every component gets a copy of the grid now. Provided gridsim builders to make it easier.
