+++ abmarl/sim/gridworld/base.py
@@ -1,7 +1,103 @@
 
 from abc import ABC
 
+import numpy as np
+
+from abmarl.sim import AgentBasedSimulation
 from abmarl.sim.gridworld.agent import GridWorldAgent
+
+
+class GridWorldSimulation(AgentBasedSimulation, ABC):
+    """
+    GridWorldSimulation interface.
+
+    Extends the AgentBasedSimulation interface for the GridWorld. We provide builders
+    for streamlining the building process. The interface is meant to be used like
+    so:
+    MyCustomGridSim(GridWorldSimulation):
+        ... # define the simulation, don't overwrite the builders.
+    
+    sim = MyCustomGridSim.build(parameters)
+    """
+    @classmethod
+    def build_sim(cls, rows, cols, **kwargs):
+        """
+        Build a GridSimulation.
+
+        Specify the number of row, the number of cols, a dictionary of agents,
+        and any additional parameters.
+
+        Args:
+            rows: The number of rows in the grid. Must be a positive integer.
+            cols: The number of cols in the grid. Must be a positive integer.
+            agents: The dictionary of agents in the grid.
+
+        Returns:
+            A GridSimulation configured as specified.
+        """
+        assert type(rows) is int, "Rows must be an integer."
+        assert 0 < rows, "Rows must be a positive integer."
+        assert type(cols) is int, "Cols must be an integer."
+        assert 0 < cols, "Cols must be a positive integer."
+
+        return cls._build_sim(rows, cols, **kwargs)
+    
+    @classmethod
+    def build_sim_from_file(cls, file_name, object_registry, **kwargs):
+        """
+        Build a GridSimulation from a text file.
+
+        Args:
+            file_name: Name of the file that specifies the initial grid setup. In the file, each
+                cell should be a single alphanumeric character indicating which agent
+                will be at that position (from the perspective of looking down on the
+                grid). That agent will be given that initial position. 0's are reserved for empty
+                space. For example:
+                0 A W W W A 0
+                0 0 0 0 0 0 0
+                0 A W W W A 0
+                will create a 3-by-7 grid with some agents along the top and bottom
+                of the grid and another type of agent in the corners.
+            object_registry: A dictionary that maps characters from the file to a
+                function that generates the agent. This must be a function because
+                each agent must have unique id, which is generated here. For example,
+                using the grid above and some pre-created Agent classes:
+                {
+                    'A': lambda n: ExploringAgent(id=f'explorer{n}', view_range=3, move_range=1),
+                    'W': lambda n: WallAgent(id=f'wall{n}')
+                }
+                will create a grid with ExploringAgents in the corners and WallAgents
+                along the top and bottom rows.
+
+        Returns:
+            A GridSimulation built from the file.
+        """
+        assert type(file_name) is str, "The file_name must be the name of the file."
+        assert type(object_registry) is dict, "The object_registry must be a dictionary."
+        assert 0 not in object_registry, "0 is reserved for empty space."
+        agents = {}
+        n = 0
+        with open(file_name, 'r') as fp:
+            lines = fp.read().splitlines()
+            cols = len(lines[0].split(' '))
+            rows = len(lines)
+            for row, line in enumerate(lines):
+                chars = line.split(' ')
+                assert len(chars) == cols, f"Mismatched number of columns per row in {file_name}"
+                for col, char in enumerate(chars):
+                    if char in object_registry:
+                        agent = object_registry[char](n)
+                        agent.initial_position = np.array([row, col])
+                        agents[agent.id] = agent
+                        n += 1
+        
[…]
