### 2021-06-17 f7843f3
Fixed assertion warning message in agent class.
