+++ abmarl/sim/gridworld/agent.py
@@ -24,7 +24,7 @@ class GridWorldAgent(PrincipleAgent):
     def initial_position(self, value):
         if value is not None:
             assert type(value) is np.ndarray, "Initial position must be a numpy array."
-            assert value.shape == (2,), "Initial position must be a 2-dimensional array."
+            assert value.shape == (2,), "Initial position must be a 2-element array."
             assert value.dtype in [np.int, np.float], "Initial position must be numerical."
         self._initial_position = value
 
