### 2021-06-17 b8baa24
PositionState resets the grid by filling it with Nones.
