### 2021-06-17 3c07e23
Cleaned up components and examples to work with top level changes
