### 2021-06-17 3aaf339
Fixed linting issues
