### 2021-06-17 21e1f87
Added reset to the grid example to ensure that grid is shared among components.
