+++ abmarl/sim/gridworld/gridworld_example.py
@@ -124,6 +124,9 @@ if __name__ == "__main__":
         }
         sim.step(action)
         sim.render(fig=fig)
+    
+    sim.reset()
+    sim.render(fig=fig)
 
     # Examine the agents' observations
     from pprint import pprint
