### 2021-06-21 bba2c98
Added configuration properties to agent classes
