+++ abmarl/sim/gridworld/agent.py
@@ -158,3 +158,29 @@ class HealthAgent(GridWorldAgent):
         """
         return self.health > 0
 
+
+class TeamAgent(GridWorldAgent):
+    """
+    Agents with team assignment.
+
+    The team value can be used to set rules for inter-agent interaction.
+    """
+    def __init__(self, team=None, **kwargs):
+        super().__init__(**kwargs)
+        self.team = team
+
+    @property
+    def team(self):
+        """
+        The agent's team must be an integer.
+        """
+        return self._team
+
+    @team.setter
+    def team(self, value):
+        assert type(value) is int, "Team must be an integer."
+        self._team = value
+
+    @property
+    def configured(self):
+        return super().configured and self.team is not None
