+++ abmarl/sim/components/agent.py
@@ -7,160 +7,11 @@ from abmarl.sim import PrincipleAgent, ActingAgent, ObservingAgent
 # --- Base Agent --- #
 # ------------------ #
 
-class ComponentAgent(PrincipleAgent):
-    """
-    Component Agents have a position, life, and team.
-
-    initial_position (np.array or None):
-        The desired starting position for this agent.
-
-    min_health (float):
-        The minimum health the agent can reach before it dies.
-        Default 0.0.
-
-    max_health (float):
-        The maximum health the agent can reach.
-        Default 1.0.
-
-    initial_health (float or None):
-        The initial health of the agent. The health will be set to this initial
-        option at reset time.
-
-    team (int or None):
-        The agent's team. Teams are indexed starting from 1, with team 0 reserved
-        for agents that are not on a team (None).
-    """
-    def __init__(self, initial_position=None, min_health=0.0, max_health=1.0, initial_health=None,
-                 team=None, **kwargs):
-        super().__init__(**kwargs)
-        self.initial_position = initial_position
-        assert type(min_health) in [float, int] and type(max_health) in [float, int], \
-            "Min and max health must be numerical."
-        assert min_health <= max_health, \
-            "The min health must be less than or equal to the max_health."
-        self._min_max_health = np.array([min_health, max_health])
-        self.initial_health = initial_health
-        self.is_alive = True
-        self.team = team
-
-    @property
-    def initial_position(self):
-        return self._initial_position
-
-    @initial_position.setter
-    def initial_position(self, value):
-        if value is not None:
-            assert type(value) is np.ndarray, "Initial position must be a numpy array."
-            assert value.shape == (2,), "Initial position must be a 2-element array."
-            assert value.dtype in [np.int, np.float], "Initial position must be numerical."
-        self._initial_position = value
-
-    @property
-    def min_health(self):
-        return self._min_max_health[0]
-
-    @property
-    def max_health(self):
-        return self._min_max_health[1]
-
-    @property
-    def initial_health(self):
-        return self._initial_health
-
-    @initial_health.setter
-    def initial_health(self, value):
-        if value is not None:
-            assert type(value) in [float, int], "Initial health must be a float."
-            assert self.min_health <= value <= self.max_health, \
-                "Initial health must be between the min and max health."
-        self._initial_health = value
-
-    @property
-    def team(self):
-        return self._team
-
-    @team.setter
-    def team(self, value):
-        if value is not None:
-            assert type(value) is int, "Team must be an int."
-            assert value != 0, "Team 0 is reserved for agents who do not have a team. " + \
-                "Use a team number greater than 0."
-            self._team = value
-        else:
-            self._team = 0
-
-    @property
-    def configured(self):
-        """
-        Determine if the agent has been successfully configured.
-        """
-        return super().configured and \
-            self._min_max_health is not None and \
-            self.is_alive is not None and \
-            self.team is not None
-
-
-# ----------------- #
-# --- Attacking --- #
-# ----------------- #
-
-class AttackingAgent(ActingAgent, ComponentAgent):
-    """
-    Agents that can attack other agents.
-
-    attack_range (int):
-        The effective range of the attack. Can be used to determine if an attack
-        is successful based on distance between agents.
-
-    attack_strength (float):
-        How effective the agent's attack is. This is applicable in situations where
-        the agents' health is affected by attacks.
-
-    attack_accuracy (float):
-        The effective accuracy of the agent's attack. Should be between 0 and 1.
-        To make deterministic attacks, use 1. Default is 1.
-    """
-    def __init__(self, attack_range=None, attack_strength=None, attack_accuracy=1, **kwargs):
-        super().__init__(**kwargs)
[…]
