### 2021-06-21 6f4aca7
Cleaned up grid observing agent
