### 2021-06-21 823c8ad
Cleaned up moving agent
