### 2021-06-21 f1a4226
Removed unnecessary classes from components/agent
