### 2021-06-21 513a87e
Added attacking actor to gridworld
