+++ abmarl/sim/gridworld/actor.py
@@ -5,7 +5,8 @@ import numpy as np
 from gym.spaces import Box
 
 from abmarl.sim.gridworld.base import GridWorldBaseComponent
-from abmarl.sim.gridworld.agent import MovingAgent
+from abmarl.sim.gridworld.state import HealthState
+from abmarl.sim.gridworld.agent import MovingAgent, AttackingAgent, HealthAgent
 
 
 class ActorBaseComponent(GridWorldBaseComponent, ABC):
@@ -95,3 +96,154 @@ class MoveActor(ActorBaseComponent):
                 self.grid[agent.position[0], agent.position[1]] = None
                 agent.position = new_position
                 self.grid[agent.position[0], agent.position[1]] = agent
+
+class AttackActor(ActorBaseComponent):
+    """
+    Agents can attack other agents.
+
+    The other agent must be within the attack_range of the attacking agent. If
+    there are multiple attackable agents in the range, then one will be randomly
+    chosen. Attackable agents are determiend by the team_matrix. The effectiveness
+    of the attack is determined by the attacking agent's strength and accuracy.
+    """
+    def __init__(self, health_state=None, attack_norm=np.inf, team_attack_matrix=None,
+            number_of_teams=None, **kwargs):
+        super().__init__(**kwargs)
+        self.health_state = health_state
+        self.attack_norm = attack_norm
+        self.number_of_teams = number_of_teams
+        self.team_attack_matrix = team_attack_matrix
+
+    @property
+    def health_state(self):
+        """
+        HealthState component that manages the state of the agents' healths.
+        """
+        return self._health_state
+
+    @health_state.setter
+    def health_state(self, value):
+        assert isinstance(value, HealthState), "Health state must be a HealthState object."
+        self._health_state = value
+
+    @property
+    def attack_norm(self):
+        """
+        Norm used to measure the distance between agents.
+
+        Attack norm must be a positive integer or np.inf.
+        """
+        return self._attack_norm
+
+    @attack_norm.setter
+    def attack_norm(self, value):
+        assert (type(value) is int and 0 < value) or value == np.inf, \
+            "Attack norm must be a positive integer or np.inf"
+        self._attack_norm = value
+
+    @property
+    def number_of_teams(self):
+        """
+        The number of teams in the simulation.
+
+        The default is 0, which indicates that there are no teams, so attacking
+        is "free-for-all".
+        """
+        return self._number_of_teams
+
+    @number_of_teams.setter
+    def number_of_teams(self, value):
+        assert type(value) is int and 0 <= value, "Number of teams must be a nonnegative integer."
+        self._number_of_teams = value
+
+    @property
+    def team_attack_matrix(self):
+        """
+        A matrix that indicates which teams can attack which other team.
+        
+        We use the value at the index, like so:
+            team_matrix[attacking_team, attacked_team] = 0 or 1.
+        0 indicates it cannot attack, 1 indicates that it can.
+
+        If this is not specified, we build one that allows agents to attack any
+        other team, including agents who do not have a team.
+        """
+        return self._team_attack_matrix
+
+    @team_attack_matrix.setter
+    def team_attack_matrix(self, value):
+        if value is None:
+            self._team_attack_matrix = -np.diag(np.ones(self.number_of_teams+1)) + 1
+            self._team_attack_matrix[0, 0] = 1
+        else:
+            assert type(value) is np.ndarray, "Team attack matrix must be a numpy array."
+            assert value.shape == (self.number_of_teams, self._number_of_teams), \
+                "Team attack matrix size must be (number_of_teams x number_of_teams.)"
+            self._team_attack_matrix = value
+
+    @property
+    def key(self):
+        """
+        This Actor's key is "attack".
[…]
