+++ abmarl/sim/components/actor.py
@@ -8,136 +8,6 @@ from abmarl.sim.components.agent import AttackingAgent, GridMovementAgent, Harve
     CollisionAgent, BroadcastingAgent
 
 
-class Actor(ABC):
-    """
-    Base actor class provides the interface required of all actors. Setup the agents'
-    action space according to the Actor's channel.
-
-        agents (dict):
-            The dictionary of agents.
-
-        instance (Agent):
-            An Agent class. This is used in the isinstance check to determine if
-            the agent will receive the action channel.
-
-        space_func (function):
-            A function that takes the agent as input and outputs the action space.
-    """
-    def __init__(self, agents=None, instance=None, space_func=None, **kwargs):
-        self.agents = agents
-        for agent in self.agents.values():
-            if isinstance(agent, instance):
-                agent.action_space[self.channel] = space_func(agent)
-
-    def _get_action_from_dict(self, action_dict, **kwargs):
-        """
-        The action dict passed to an AgentBasedSimulation will be keyed off the
-        agent id's. This action dict, however, is just the specific agent's actions,
-        which is a dictionary keyed off the action channels. This function extracts
-        that action if available, otherwise returning the Actor's null_value.
-        """
-        return action_dict.get(self.channel, self.null_value)
-
-    @abstractmethod
-    def process_action(self, agent, action_dict, **kwargs):
-        pass
-
-    @abstractproperty
-    def channel(self):
-        pass
-
-    @abstractproperty
-    def null_value(self):
-        pass
-
-
-# ----------------- #
-# --- Attacking --- #
-# ----------------- #
-
-class AttackActor(Actor):
-    """
-    Agents can attack other agents within their attack radius. If there are multiple
-    attackable agents in the radius, then one will be randomly chosen. Attackable
-    agents are determiend by the team_matrix.
-
-    agents (dict of Agents):
-        The dictionary of agents.
-
-    attack_norm (int):
-        Norm used to measure the distance between agents. For example, you might
-        use a norm of 1 or np.inf in a Gird space, while 2 might be used in a Continuous
-        space. Default is np.inf.
-
-    team_attack_matrix (np.ndarray):
-        A matrix that indicates which teams can attack which other team using the
-        value at the index, like so:
-            team_matrix[attacking_team, attacked_team] = 0 or 1.
-        0 indicates it cannot attack, 1 indicates that it can.
-        Default None, meaning that any team can attack any other team, and no team
-        can attack itself.
-
-    number_of_teams (int):
-        Specify the number of teams in the simulation for building the team_attack_matrix
-        if that is not specified here.
-        Default 0, indicating that there are no teams and its a free-for-all battle.
-    """
-    def __init__(self, attack_norm=np.inf, team_attack_matrix=None, number_of_teams=0, **kwargs):
-        super().__init__(
-            instance=AttackingAgent,
-            space_func=lambda agent: Discrete(2),
-            **kwargs
-        )
-        if team_attack_matrix is None:
-            # Default: teams can attack all other teams but not themselves. Agents
-            # that are "on team 0" are actually teamless, so they can be attacked
-            # by and can attack agents from any other team, including "team 0"
-            # agents.
-            self.team_attack_matrix = -np.diag(np.ones(number_of_teams+1)) + 1
-            self.team_attack_matrix[0, 0] = 1
-        else:
-            self.team_attack_matrix = team_attack_matrix
-        self.attack_norm = attack_norm
-
-    def process_action(self, attacking_agent, action_dict, **kwargs):
-        """
-        If the agent has chosen to attack, then determine which agent got attacked.
-
-        return (Agent):
-            Return the attacked agent object. This can be None if no agent was
-            attacked.
-        """
-        if self._get_action_from_dict(action_dict):
-            for attacked_agent in self.agents.values():
[…]
