### 2021-06-21 c525f2f
Removed unneeded actor components
