+++ abmarl/sim/components/actor.py
@@ -1,318 +1,0 @@
-from abc import ABC, abstractmethod, abstractproperty
-
-from gym.spaces import Discrete, Box
-import numpy as np
-
-from abmarl.sim.components.agent import AttackingAgent, GridMovementAgent, HarvestingAgent, \
-    SpeedAngleAgent, AcceleratingAgent, VelocityAgent, \
-    CollisionAgent, BroadcastingAgent
-
-
-
-
-# ----------------------------- #
-# --- Position and Movement --- #
-# ----------------------------- #
-
-class GridMovementActor(Actor):
-    """
-    Provides the necessary action space for agents who can move and processes such
-    movements.
-
-    position (GridPositionState):
-        The position state handler. Needed to modify the agents' positions.
-
-    agents (dict):
-        The dictionary of agents.
-    """
-    def __init__(self, position_state=None, **kwargs):
-        super().__init__(
-            instance=GridMovementAgent,
-            space_func=lambda agent: Box(-agent.move_range, agent.move_range, (2,), np.int),
-            **kwargs
-        )
-        self.position_state = position_state
-
-    def process_action(self, agent, action_dict, **kwargs):
-        """
-        Determine the agent's new position based on its move action.
-
-        return (np.array):
-            How much the agent has moved in row and column. This can be different
-            from the desired move if the position update was invalid.
-        """
-        move = self._get_action_from_dict(action_dict)
-        position_before = agent.position
-        self.position_state.modify_position(agent, move, **kwargs)
-        return agent.position - position_before
-
-    @property
-    def channel(self):
-        return 'move'
-
-    @property
-    def null_value(self):
-        return np.zeros(2)
-
-
-class SpeedAngleMovementActor:
-    """
-    Process acceleration and angle changes for SpeedAngleAgents. Update the agents'
-    positions based on their new speed and direction.
-
-    position (ContinuousPositionState):
-        The position state handler. Needed to modify agent positions.
-
-    speed_angle (SpeedAngleState):
-        The speed and angle state handler.
-
-    agents (dict):
-        The dictionary of agents.
-    """
-    def __init__(self, position_state=None, speed_angle_state=None, agents=None, **kwargs):
-        self.position_state = position_state
-        self.speed_angle_state = speed_angle_state
-        self.agents = agents
-
-        for agent in agents.values():
-            if isinstance(agent, SpeedAngleAgent):
-                agent.action_space['accelerate'] = Box(
-                    -agent.max_acceleration, agent.max_acceleration, (1,)
-                )
-                agent.action_space['bank'] = Box(
-                    -agent.max_banking_angle_change, agent.max_banking_angle_change, (1,)
-                )
-
-    def process_move(self, agent, acceleration, angle, **kwargs):
-        """
-        Update the agent's speed by applying the acceleration and the agent's banking
-        angle by applying the change. Then use the updated speed and ground angle
-        to determine the agent's next position.
-
-        agent (SpeedAngleAgent):
-            Agent that is attempting to move.
-
-        acceleration (np.array):
-            A one-element float array that changes the agent's speed. New speed
-            must be within the agent's min and max speed.
-
-        angle (np.array):
-            A one-element float array that changes the agent's banking angle. New
-            banking angle must be within the agent's min and max banking angles.
-
-        return (np.array):
-            Return the change in position.
-        """
-        if isinstance(agent, SpeedAngleAgent):
-            self.speed_angle_state.modify_speed(agent, acceleration[0])
-            self.speed_angle_state.modify_banking_angle(agent, angle[0])
-
-            x_position = agent.speed*np.cos(np.deg2rad(agent.ground_angle))
-            y_position = agent.speed*np.sin(np.deg2rad(agent.ground_angle))
-
[…]
