### 2021-06-21 e9f50ba
Removed uncessary observers
