+++ abmarl/sim/components/observer.py
@@ -1,607 +1,0 @@
-from abc import ABC, abstractmethod, abstractproperty
-
-from gym.spaces import Box, Dict
-import numpy as np
-
-from abmarl.sim.components.agent import HealthObservingAgent, LifeObservingAgent, \
-    AgentObservingAgent, PositionObservingAgent, SpeedAngleObservingAgent, \
-    VelocityObservingAgent, ResourceObservingAgent, TeamObservingAgent, BroadcastObservingAgent, \
-    SpeedAngleAgent, VelocityAgent, BroadcastingAgent, ComponentAgent
-
-
-class Observer(ABC):
-    """
-    Base observer class provides the interface required of all observers. Setup
-    the agents' observation space according to the Observer's channel.
-
-        agents (dict):
-            The dictionary of agents.
-    """
-    def __init__(self, agents=None, **kwargs):
-        self.agents = agents
-
-    def _set_obs_space_simple(self, instance, space_func, **kwargs):
-        """
-        Observers that don't depend on the type of the other agents can use this
-        method.
-
-        instance (Agent):
-            An Agent class. This is used in the isinstance check to determine if
-            the agent will receive the observation channel.
-
-        space_func (function):
-            A function that takes the other agent as input and outputs the
-            observation space.
-        """
-        for agent in self.agents.values():
-            if isinstance(agent, instance):
-                agent.observation_space[self.channel] = Dict({
-                    other.id: space_func(other) for other in self.agents.values()
-                })
-
-    def _set_obs_space(self, instance, other_instance, space_func, alt_space_func, **kwargs):
-        """
-        Observers that depend on the type of the other agents must use this method.
-
-        instance (Agent):
-            An Agent class. This is used in the isinstance check to determine if
-            the agent will receive the observation channel.
-
-        other_instance (Agent):
-            An Agent class. This is used in the isinstance check on the other agents
-            to determine whether to use the space_func or the alt_space_func.
-
-        space_func (function):
-            A function that takes the other agent as input and outputs the
-            observation space.
-
-        alt_space_func (function):
-            Use this function for cases when the isinstance check fails on the
-            other agent. Function does not have inputs and outputs observation space.
-        """
-        for agent in self.agents.values():
-            if isinstance(agent, instance):
-                obs_space = {}
-                for other in self.agents.values():
-                    if isinstance(other, other_instance):
-                        obs_space[other.id] = space_func(other)
-                    else:
-                        obs_space[other.id] = alt_space_func()
-                agent.observation_space[self.channel] = Dict(obs_space)
-
-    def _get_obs(self, agent, instance=None, other_instance=ComponentAgent, attr=None, **kwargs):
-        """
-        Many observers just directly query the corresponding state field from the
-        agent. This function does exactly that, checking the instance of the observing
-        agent and the other agents and setting the observation value accordingly.
-        """
-        if isinstance(agent, instance):
-            obs = {}
-            for other in self.agents.values():
-                if isinstance(other, other_instance):
-                    obs[other.id] = getattr(other, attr)
-                else:
-                    obs[other.id] = self.null_value
-            return {self.channel: obs}
-        else:
-            return {}
-
-    @abstractmethod
-    def get_obs(self, agent, **kwargs):
-        pass
-
-    @abstractproperty
-    def channel(self):
-        pass
-
-    @abstractproperty
-    def null_value(self):
-        pass
-
-
-# --------------------- #
[…]
