### 2021-06-21 b40976d
Ported done component to gridworld framework
