+++ abmarl/sim/gridworld/state.py
@@ -9,7 +9,47 @@ from abmarl.sim.gridworld.base import GridWorldBaseComponent
 class StateBaseComponent(GridWorldBaseComponent, ABC):
     """
     Abstract State Component base from which all state components will inherit.
+
+    In addition to the Grid World Components, every State Component has access
+    to the grid.
     """
+    def __init__(self, grid=None, **kwargs):
+        self.grid = grid
+        self._rows = self.grid.shape[0]
+        self._cols = self.grid.shape[1]
+
+    @property
+    def rows(self):
+        """
+        The number of rows in the grid.
+        """
+        return self._rows
+
+    @property
+    def cols(self):
+        """
+        The number of columns in the grid.
+        """
+        return self._cols
+
+    @property
+    def grid(self):
+        """
+        The grid indexes the agents by their position.
+
+        For example, an agent whose position is (3, 2) can be accessed through
+        the grid with self.grid[3, 2]. Components are responsible for maintaining
+        the connection between agent position and grid index.
+        """
+        return self._grid
+
+    @grid.setter
+    def grid(self, value):
+        assert type(value) is np.ndarray, "The grid must be a numpy array."
+        assert len(value.shape) == 2, "The grid must be a 2-dimensional array."
+        assert value.dtype is np.dtype(object), "The grid must be a numpy array of objects."
+        self._grid = value
+
     @abstractmethod
     def reset(self, **kwargs):
         """
