### 2021-06-21 e8ca6fe
Position state handles the agent's new positions in the grid
