### 2021-06-21 847329d
MovingActor uses position state to manage the state
