### 2021-06-21 190887f
Fixed call to health_state from attack actor
