### 2021-06-21 bf6baf7
All agents have an active property. Is alive replaced by active
