+++ abmarl/sim/gridworld/examples/gridworld_example.py
@@ -3,9 +3,10 @@ from matplotlib import pyplot as plt
 import numpy as np
 
 from abmarl.sim.gridworld.base import GridWorldSimulation
-from abmarl.sim.gridworld.agent import GridWorldAgent, GridObservingAgent, MovingAgent
-from abmarl.sim.gridworld.state import PositionState
-from abmarl.sim.gridworld.actor import MoveActor
+from abmarl.sim.gridworld.agent import GridWorldAgent, GridObservingAgent, MovingAgent, \
+    AttackingAgent, HealthAgent, TeamAgent
+from abmarl.sim.gridworld.state import PositionState, HealthState
+from abmarl.sim.gridworld.actor import MoveActor, AttackActor
 from abmarl.sim.gridworld.observer import GridObserver
 from abmarl.tools.matplotlib_utils import mscatter
 
@@ -16,20 +17,49 @@ class WallAgent(GridWorldAgent):
 
     Args:
         encoding: Default encoding is 1.
+        render_shape: Default render_shape set to 'X'.
     """
-    def __init__(self, encoding=1, **kwargs):
+    def __init__(self, encoding=1, render_shape='X', **kwargs):
+        super().__init__(**kwargs)
+        self.encoding = encoding
+        self.render_shape = render_shape
+
+
+class FoodAgent(HealthAgent, TeamAgent):
+    """
+    Food Agents do not move and can be attacked by Foraging Agents.
+    
+    Args:
+        encoding: Default encoding set to 2.
+    """
+    def __init__(self, encoding=2, **kwargs):
         super().__init__(**kwargs)
         self.encoding = encoding
 
 
-class ExploringAgent(MovingAgent, GridObservingAgent):
+class ForagingAgent(HealthAgent, TeamAgent, AttackingAgent, MovingAgent, GridObservingAgent):
     """
-    Exploring agents, moving around and observing the grid.
+    Foraging Agents can move, attack Food agents, and be attacked by Hunting agents.
+    
+    Args:
+        encoding: Default encoding set to 3.
+        render_shape: Default render_shape set to 'o'.
+    """
+    def __init__(self, encoding=3, render_shape='o', **kwargs):
+        super().__init__(**kwargs)
+        self.encoding = encoding
+        self.render_shape = render_shape
+
+
+class HuntingAgent(HealthAgent, AttackingAgent, MovingAgent, GridObservingAgent, TeamAgent):
+    """
+    Hunting agents can move and attack Foraging agents.
 
     Args:
-        encoding: Default encoding is 2.
+        encoding: Default encoding set to 4.
+        render_shape: Default render_shape set to 'D'.
     """
-    def __init__(self, encoding=2, render_shape='o', **kwargs):
+    def __init__(self, encoding=4, render_shape='D', **kwargs):
         super().__init__(**kwargs)
         self.encoding = encoding
         self.render_shape = render_shape
@@ -41,8 +71,10 @@ class GridSim(GridWorldSimulation):
 
         # State Components
         self.grid_state = PositionState(**kwargs)
+        self.health_state = HealthState(**kwargs)
 
         # Action Components
+        self.attack_actor = AttackActor(health_state=self.health_state, **kwargs)
         self.move_actor = MoveActor(position_state=self.grid_state, **kwargs)
 
         # Observation Components
@@ -52,8 +84,16 @@ class GridSim(GridWorldSimulation):
 
     def reset(self, **kwargs):
         self.grid_state.reset(**kwargs)
+        self.health_state.reset(**kwargs)
 
     def step(self, action_dict, **kwargs):
+        # Process attacks:
+        for agent_id, action in action_dict.items():
+            agent = self.agents[agent_id]
[…]
