### 2021-06-21 6e89947
updatead build_grid_example
