+++ abmarl/sim/gridworld/examples/build_grid_example.py
@@ -1,10 +1,10 @@
 
 from matplotlib import pyplot as plt
 
-from abmarl.sim.gridworld.examples.gridworld_example import WallAgent, ExploringAgent, GridSim
+from abmarl.sim.gridworld.examples.gridworld_example import WallAgent, FoodAgent, GridSim
 
 object_registry = {
-    'A': lambda n: ExploringAgent(id=f'explorer{n}', view_range=3, move_range=1),
+    'A': lambda n: FoodAgent(id=f'explorer{n}', initial_health=1, team=1),
     'W': lambda n: WallAgent(id=f'wall{n}')
 }
 
@@ -12,7 +12,7 @@ file_name = 'starting_grid.txt'
 
 fig = plt.figure()
 
-sim = GridSim.build_sim_from_file(file_name, object_registry)
+sim = GridSim.build_sim_from_file(file_name, object_registry, number_of_teams=1)
 sim.reset()
 sim.render(fig=fig)
 
