### 2021-06-22 842ad58
Small fix ups from PR
