+++ abmarl/sim/agent_based_simulation.py
@@ -35,9 +35,9 @@ class PrincipleAgent:
         """
         True if the agent is still active in the simulation.
 
-        Active means that the agent is in a valid state. For exmaple, if agents
-        can die, then active is True if the agents are alive or False if they're
-        dead.
+        Active means that the agent is in a valid state. For example, suppose agents
+        in our Simulation can die. Then active is True if the agents are alive
+        or False if they're dead.
         """
         return True
 
