### 2021-06-22 7e2298c
Removed attack norm
