### 2021-06-22 72184c6
Grid updates given the the actors and not the state components
