### 2021-06-22 d21c2be
View blocking agents also block attacks and attack search is now random within the agent's attack_range.
