### 2021-06-22 f6338c7
Attack mapping uses agents' encoding to determine possible attacks
