### 2021-06-22 e5b9a91
Small fixes per PR
