+++ abmarl/sim/gridworld/done.py
@@ -4,7 +4,6 @@ from abc import ABC, abstractmethod
 import numpy as np
 
 from abmarl.sim.gridworld.base import GridWorldBaseComponent
-from abmarl.sim.gridworld.agent import HealthAgent
 
 class DoneBaseComponent(GridWorldBaseComponent, ABC):
     """
