### 2021-06-22 #3360 @rusu24edward
@drvinceknight I'm sorry, I did not realize that github was sensitive to capitalization in the branch name. The name of the branch is `abmarl-87-interface-release`. I've attempted compilation just now, and received the same error as before. This is strange to me because the pdf will compile with the GitHub action... I'll look into resolving this issue, any tips?
