### 2021-06-22 #3360 @rusu24edward
@drvinceknight I believe I've fixed the pdf compilation issue. I had forgotten to add an `index` field under the affiliation. It looks like the paper has compiled correctly now.
