### 2021-06-23 #3360 @drvinceknight
@satpreetsingh @seba-1511  @abhiramm7 @bozutemiz 'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “Abmarl: Connecting Agent-Based Simulations with Multi-Agent Reinforcement Learning". You can find more information at the top of this Github issue.
