### 2021-06-25 51a1fe0
Added nonoverlapping and overlapping grids
