### 2021-06-25 4c10708
Position State enforces nonoverlapping grid
