### 2021-06-25 beefdff
Position state to UniquePosition state
