### 2021-06-25 d3b0765
Unique position uses grid api. Added overlapping position state
