### 2021-06-25 fb6bba5
Overlappable grids use set intead of list
