### 2021-06-25 c20fb05
Removed position state from move actor becuase it does not need it
