+++ abmarl/sim/gridworld/actor.py
@@ -55,26 +55,13 @@ class MoveActor(ActorBaseComponent):
     """
     Agents can move to unoccupied nearby squares.
     """
-    def __init__(self, position_state=None, **kwargs):
+    def __init__(self, **kwargs):
         super().__init__(**kwargs)
-        self.position_state = position_state
         for agent in self.agents.values():
             if isinstance(agent, self.supported_agent_type):
                 agent.action_space[self.key] = Box(
                     -agent.move_range, agent.move_range, (2,), np.int
                 )
-
-    @property
-    def position_state(self):
-        """
-        PositionState component that manages the state of the agents' positions.
-        """
-        return self._position_state
-
-    @position_state.setter
-    def position_state(self, value):
-        assert isinstance(value, UniquePositionState), "Position state must be a PositionState object."
-        self._position_state = value
 
     @property
     def key(self):
