### 2021-06-25 b388393
Added query function to grid api
