### 2021-06-25 bc12b90
Overlappable Position handler
