### 2021-06-25 71e3179
tuple index on query and remove
