### 2021-06-25 a35f5d4
Agents with overlappable property
