### 2021-06-25 5acfbb7
Resovling data structures issues in the overlapping grid
