### 2021-06-28 #3424 @whedon
[REVIEW]: Abmarl: Connecting Agent-Based Simulations with Multi-Agent Reinforcement Learning
**Submitting author:** @rusu24edward (<a href="http://orcid.org/0000-0003-1033-439X">Edward Rusu</a>)
**Repository:** <a href="https://github.com/LLNL/Abmarl/" target ="_blank">https://github.com/LLNL/Abmarl/</a>
**Version:** 0.1.4

- [x] #3515
