### 2021-06-28 #3424 @drvinceknight
> MISSING DOIs
>
> - 10.1613/jair.3912 may be a valid DOI for title: The Arcade Learning Environment: An Evaluation Platform for General Agents

@rusu24edward if you could take a look at the above missing DOI please.
