+++ publications/paper.bib
@@ -28,6 +28,7 @@
     month = "jun",
     volume = "47",
     pages = "253--279",
+    doi = {10.1613/jair.3912},
 }
 
 @misc{hide-n-seek,
