+++ abmarl/sim/gridworld/maze.py
@@ -1,0 +1,257 @@
+# Maze generator -- Randomized Prim Algorithm
+
+## Imports
+import random
+import time
+from colorama import init
+from colorama import Fore, Back, Style
+
+## Functions
+def printMaze(maze):
+	for i in range(0, height):
+		for j in range(0, width):
+			if (maze[i][j] == 'u'):
+				print(Fore.WHITE + str(maze[i][j]), end=" ")
+			elif (maze[i][j] == 'c'):
+				print(Fore.GREEN + str(maze[i][j]), end=" ")
+			else:
+				print(Fore.RED + str(maze[i][j]), end=" ")
+			
+		print('\n')
+
+# Find number of surrounding cells
+def surroundingCells(rand_wall):
+	s_cells = 0
+	if (maze[rand_wall[0]-1][rand_wall[1]] == 'c'):
+		s_cells += 1
+	if (maze[rand_wall[0]+1][rand_wall[1]] == 'c'):
+		s_cells += 1
+	if (maze[rand_wall[0]][rand_wall[1]-1] == 'c'):
+		s_cells +=1
+	if (maze[rand_wall[0]][rand_wall[1]+1] == 'c'):
+		s_cells += 1
+
+	return s_cells
+
+
+## Main code
+# Init variables
+wall = 'w'
+cell = 'c'
+unvisited = 'u'
+height = 11
+width = 27
+maze = []
+
+# Initialize colorama
+init()
+
+# Denote all cells as unvisited
+for i in range(0, height):
+	line = []
+	for j in range(0, width):
+		line.append(unvisited)
+	maze.append(line)
+
+# Randomize starting point and set it a cell
+starting_height = int(random.random()*height)
+starting_width = int(random.random()*width)
+if (starting_height == 0):
+	starting_height += 1
+if (starting_height == height-1):
+	starting_height -= 1
+if (starting_width == 0):
+	starting_width += 1
+if (starting_width == width-1):
+	starting_width -= 1
+
+# Mark it as cell and add surrounding walls to the list
+maze[starting_height][starting_width] = cell
+walls = []
+walls.append([starting_height - 1, starting_width])
+walls.append([starting_height, starting_width - 1])
+walls.append([starting_height, starting_width + 1])
+walls.append([starting_height + 1, starting_width])
+
+# Denote walls in maze
+maze[starting_height-1][starting_width] = 'w'
+maze[starting_height][starting_width - 1] = 'w'
+maze[starting_height][starting_width + 1] = 'w'
+maze[starting_height + 1][starting_width] = 'w'
+
+while (walls):
+	# Pick a random wall
+	rand_wall = walls[int(random.random()*len(walls))-1]
+
+	# Check if it is a left wall
+	if (rand_wall[1] != 0):
+		if (maze[rand_wall[0]][rand_wall[1]-1] == 'u' and maze[rand_wall[0]][rand_wall[1]+1] == 'c'):
+			# Find the number of surrounding cells
+			s_cells = surroundingCells(rand_wall)
+
+			if (s_cells < 2):
+				# Denote the new path
+				maze[rand_wall[0]][rand_wall[1]] = 'c'
+
+				# Mark the new walls
+				# Upper cell
+				if (rand_wall[0] != 0):
+					if (maze[rand_wall[0]-1][rand_wall[1]] != 'c'):
+						maze[rand_wall[0]-1][rand_wall[1]] = 'w'
[…]
