### 2021-07-08 f99d3d3
Maze generation is a function
