### 2021-07-08 8b02633
Changed generate maze input args
