+++ abmarl/sim/gridworld/maze.py
@@ -5,14 +5,11 @@ import random
 
 
 
-def generate_maze(height=11, width=27, wall='w', agents=None):
+def generate_maze(rows, cols, wall='w', target='T', agents=None):
     ## Main code
     # Init variables
-    wall = 'w'
     cell = 'c'
     unvisited = 'u'
-    height = 11
-    width = 27
     maze = []
 
     # Find number of surrounding cells
@@ -29,22 +26,22 @@ def generate_maze(height=11, width=27, wall='w', agents=None):
         return s_cells
 
     # Denote all cells as unvisited
-    for i in range(0, height):
+    for i in range(0, rows):
         line = []
-        for j in range(0, width):
+        for j in range(0, cols):
             line.append(unvisited)
         maze.append(line)
 
     # Randomize starting point and set it a cell
-    starting_height = int(random.random()*height)
-    starting_width = int(random.random()*width)
+    starting_height = int(random.random()*rows)
+    starting_width = int(random.random()*cols)
     if (starting_height == 0):
         starting_height += 1
-    if (starting_height == height-1):
+    if (starting_height == rows-1):
         starting_height -= 1
     if (starting_width == 0):
         starting_width += 1
-    if (starting_width == width-1):
+    if (starting_width == cols-1):
         starting_width -= 1
 
     # Mark it as cell and add surrounding walls to the list
@@ -85,7 +82,7 @@ def generate_maze(height=11, width=27, wall='w', agents=None):
 
 
                     # Bottom cell
-                    if (rand_wall[0] != height-1):
+                    if (rand_wall[0] != rows-1):
                         if (maze[rand_wall[0]+1][rand_wall[1]] != 'c'):
                             maze[rand_wall[0]+1][rand_wall[1]] = 'w'
                         if ([rand_wall[0]+1, rand_wall[1]] not in walls):
@@ -131,7 +128,7 @@ def generate_maze(height=11, width=27, wall='w', agents=None):
                             walls.append([rand_wall[0], rand_wall[1]-1])
 
                     # Rightmost cell
-                    if (rand_wall[1] != width-1):
+                    if (rand_wall[1] != cols-1):
                         if (maze[rand_wall[0]][rand_wall[1]+1] != 'c'):
                             maze[rand_wall[0]][rand_wall[1]+1] = 'w'
                         if ([rand_wall[0], rand_wall[1]+1] not in walls):
@@ -145,7 +142,7 @@ def generate_maze(height=11, width=27, wall='w', agents=None):
                 continue
 
         # Check the bottom wall
-        if (rand_wall[0] != height-1):
+        if (rand_wall[0] != rows-1):
             if (maze[rand_wall[0]+1][rand_wall[1]] == 'u' and maze[rand_wall[0]-1][rand_wall[1]] == 'c'):
 
                 s_cells = surroundingCells(rand_wall)
@@ -154,7 +151,7 @@ def generate_maze(height=11, width=27, wall='w', agents=None):
                     maze[rand_wall[0]][rand_wall[1]] = 'c'
 
                     # Mark the new walls
-                    if (rand_wall[0] != height-1):
+                    if (rand_wall[0] != rows-1):
                         if (maze[rand_wall[0]+1][rand_wall[1]] != 'c'):
                             maze[rand_wall[0]+1][rand_wall[1]] = 'w'
                         if ([rand_wall[0]+1, rand_wall[1]] not in walls):
@@ -164,7 +161,7 @@ def generate_maze(height=11, width=27, wall='w', agents=None):
                             maze[rand_wall[0]][rand_wall[1]-1] = 'w'
[…]
