+++ abmarl/sim/gridworld/maze.py
@@ -15,13 +15,13 @@ def generate_maze(rows, cols, wall='w', target='T', agents=None):
     # Find number of surrounding cells
     def surroundingCells(rand_wall):
         s_cells = 0
-        if (maze[rand_wall[0]-1][rand_wall[1]] == 'c'):
+        if (maze[rand_wall[0]-1][rand_wall[1]] == cell):
             s_cells += 1
-        if (maze[rand_wall[0]+1][rand_wall[1]] == 'c'):
+        if (maze[rand_wall[0]+1][rand_wall[1]] == cell):
             s_cells += 1
-        if (maze[rand_wall[0]][rand_wall[1]-1] == 'c'):
+        if (maze[rand_wall[0]][rand_wall[1]-1] == cell):
             s_cells +=1
-        if (maze[rand_wall[0]][rand_wall[1]+1] == 'c'):
+        if (maze[rand_wall[0]][rand_wall[1]+1] == cell):
             s_cells += 1
         return s_cells
 
@@ -64,18 +64,18 @@ def generate_maze(rows, cols, wall='w', target='T', agents=None):
 
         # Check if it is a left wall
         if (rand_wall[1] != 0):
-            if (maze[rand_wall[0]][rand_wall[1]-1] == 'u' and maze[rand_wall[0]][rand_wall[1]+1] == 'c'):
+            if (maze[rand_wall[0]][rand_wall[1]-1] == 'u' and maze[rand_wall[0]][rand_wall[1]+1] == cell):
                 # Find the number of surrounding cells
                 s_cells = surroundingCells(rand_wall)
 
                 if (s_cells < 2):
                     # Denote the new path
-                    maze[rand_wall[0]][rand_wall[1]] = 'c'
+                    maze[rand_wall[0]][rand_wall[1]] = cell
 
                     # Mark the new walls
                     # Upper cell
                     if (rand_wall[0] != 0):
-                        if (maze[rand_wall[0]-1][rand_wall[1]] != 'c'):
+                        if (maze[rand_wall[0]-1][rand_wall[1]] != cell):
                             maze[rand_wall[0]-1][rand_wall[1]] = 'w'
                         if ([rand_wall[0]-1, rand_wall[1]] not in walls):
                             walls.append([rand_wall[0]-1, rand_wall[1]])
@@ -83,14 +83,14 @@ def generate_maze(rows, cols, wall='w', target='T', agents=None):
 
                     # Bottom cell
                     if (rand_wall[0] != rows-1):
-                        if (maze[rand_wall[0]+1][rand_wall[1]] != 'c'):
+                        if (maze[rand_wall[0]+1][rand_wall[1]] != cell):
                             maze[rand_wall[0]+1][rand_wall[1]] = 'w'
                         if ([rand_wall[0]+1, rand_wall[1]] not in walls):
                             walls.append([rand_wall[0]+1, rand_wall[1]])
 
                     # Leftmost cell
                     if (rand_wall[1] != 0):	
-                        if (maze[rand_wall[0]][rand_wall[1]-1] != 'c'):
+                        if (maze[rand_wall[0]][rand_wall[1]-1] != cell):
                             maze[rand_wall[0]][rand_wall[1]-1] = 'w'
                         if ([rand_wall[0], rand_wall[1]-1] not in walls):
                             walls.append([rand_wall[0], rand_wall[1]-1])
@@ -105,31 +105,31 @@ def generate_maze(rows, cols, wall='w', target='T', agents=None):
 
         # Check if it is an upper wall
         if (rand_wall[0] != 0):
-            if (maze[rand_wall[0]-1][rand_wall[1]] == 'u' and maze[rand_wall[0]+1][rand_wall[1]] == 'c'):
[…]
