### 2021-07-08 0a87386
Removed entrace and exit
