### 2021-07-08 1e9e2a6
Replaced 'w' with wall_char
