### 2021-07-08 fcbde2f
wall is W and cell is 0
