### 2021-07-08 72cf05c
Add agents to the maze at empty cells
