### 2021-07-08 cad3335
Maze rows and cols fixed
