### 2021-07-08 5d3a381
Added hunting and foraging demo
