+++ examples/hunting_foraging_demo.py
@@ -1,0 +1,191 @@
+
+# ----------------------------- #
+# --- Setup the environment --- #
+# ----------------------------- #
+
+# --- Create the agents and the environment --- #
+
+# Import the simulation environment and agents
+# from admiral.envs.components.examples.hunting_and_foraging import HuntingForagingEnv, HuntingForagingAgent, FoodAgent
+from admiral.envs.simple_env import HuntingForagingEnv, HuntingForagingAgent, FoodAgent
+
+# Instatiate the agents that will operate in this environment. All possible agent
+# attributes are listed below.
+
+# Food agents are not really agents in the RL sense. They're just entites for the
+# foragers to eat.
+food = {f'food{i}': FoodAgent(id=f'food{i}', team=1) for i in range(12)}
+
+# Foragers try to eat all the food agents before they die
+foragers = {f'forager{i}': HuntingForagingAgent(
+    id=f'forager{i}',
+    agent_view=3, # Partial Observation Mask: how far away this agent can see other agents.
+    team=2, # Which team this agent is on
+    move_range=1, # How far the agent can move within a single step.
+    min_health=0.0, # If the agent's health falls below this value, it will die.
+    max_health=1.0, # Agent's health cannot grow above this value.
+    attack_range=1, # How far this agent's attack will reach.
+    attack_strength=1.0, # How powerful the agent's attack is.
+    attack_accuracy=1.0, # Probability of successful attack
+    initial_position=None # Episode-starting position. If None, then random within the region.
+) for i in range(5)}
+
+# # Hunters try to eat all the foraging agents before all the food disappears.
+hunters =  {f'hunter{i}': HuntingForagingAgent(
+    id=f'hunter{i}', 
+    agent_view=2, # Partial Observation Mask: how far away this agent can see other agents.
+    team=3, # Which team this agent is on
+    move_range=1, # How far the agent can move within a single step.
+    min_health=0.0, # If the agent's health falls below this value, it will die.
+    max_health=1.0, # Agent's health cannot grow above this value.
+    attack_range=1, # How far this agent's attack will reach.
+    attack_strength=1.0, # How powerful the agent's attack is.
+    attack_accuracy=1.0, # Probability of successful attack
+    initial_position=None # Episode-starting position. If None, then random within the region.
+) for i in range(2)}
+
+agents = {**food, **foragers, **hunters}
+
+# Instantiate the environment
+
+# Set the size of the map
+region = 20
+
+# Determine which teams can "attack" each other. In this scenario, team 2 is the
+# foragers, and they can attack the food, which is team 1. Team 3 is the hunters
+# and they can attack the foragers. So we setup a matrix that represents this.
+import numpy as np
+team_attack_matrix = np.zeros((4, 4))
+team_attack_matrix[2, 1] = 1 # Foragers can attack food
+team_attack_matrix[3, 2] = 1 # Hunters can attack foragers
+env = HuntingForagingEnv(
+    region=region, # The size of the region, both x and y
+    number_of_teams=3, # The number of teams
+    agents=agents, # Give the environment the dictionary of agents we created above
+    team_attack_matrix=team_attack_matrix,
+    # attack_norm=np.inf, # The norm to use. Default is np.inf, which means that the attack radius is square box around the agent
+)
+
+# Hack the action space
+from gym.spaces import Discrete
+for agent in agents.values():
+    agent.action_space = Discrete(10)
+
+# --- Prepare the environment for use with RLlib --- #
+
+# Now that you've created the environment, you must wrap it with a simulation manager,
+# which controls the timing of the simulation step.
[…]
