### 2021-07-08 a8594f8
Ported hunting and foraging simulation
