+++ admiral/envs/components/examples/hunting_and_foraging.py
@@ -1,0 +1,189 @@
+
+from matplotlib import pyplot as plt
+import numpy as np
+
+# Import all the features that we need from the simulation components
+from admiral.envs.components.state import GridPositionState, LifeState
+from admiral.envs.components.observer import PositionObserver, LifeObserver, TeamObserver, GridPositionBasedObserver
+from admiral.envs.components.wrappers.observer_wrapper import PositionRestrictedObservationWrapper
+from admiral.envs.components.actor import GridMovementActor, AttackActor
+from admiral.envs.components.done import AnyTeamDeadDone, TeamDeadDone
+
+# Environment needs a corresponding agent component
+from admiral.envs.components.agent import TeamAgent, PositionAgent, LifeAgent, AttackingAgent, GridMovementAgent, AgentObservingAgent, PositionObservingAgent, TeamObservingAgent, LifeObservingAgent
+
+# Import the interface
+from admiral.envs import AgentBasedSimulation
+
+# Import extra tools
+from admiral.tools.matplotlib_utils import mscatter
+
+# All HuntingForagingAgents
+# have a position, team, and life/death state
+# can observe positions, teams, and life state of other agents
+# can move around the grid and attack other agents
+class HuntingForagingAgent(TeamAgent, PositionAgent, LifeAgent, AttackingAgent, GridMovementAgent, AgentObservingAgent, PositionObservingAgent, TeamObservingAgent, LifeObservingAgent): pass
+
+# All FoodAgents
+# have a tema, position, and life
+# They are not really "agents" in the RL sense, they're just entities in the simulation for the foragers to gather.
+class FoodAgent(TeamAgent, PositionAgent, LifeAgent): pass
+
+# Create the simulation environment from the components
+class HuntingForagingEnv(AgentBasedSimulation):
+    def __init__(self, **kwargs):
+        # Explicitly pull out the the dictionary of agents. This makes the env
+        # easier to work with.
+        self.agents = kwargs['agents']
+
+        # State components
+        # These components track the state of the agents. This environment supports
+        # agents with positions, life, and team.
+        self.position_state = GridPositionState(**kwargs)
+        self.life_state = LifeState(**kwargs)
+
+        # Observer components
+        # These components handle the observations that the agents receive whenever
+        # get_obs is called. In this environment supports agents that can observe
+        # the position, health, and team of other agents and itself.
+        self.partial_observer = GridPositionBasedObserver(position=self.position_state, **kwargs)
+        # position_observer = PositionObserver(position=self.position_state, **kwargs)
+        # team_observer = TeamObserver(**kwargs)
+        # life_observer = LifeObserver(**kwargs)
+        # self.partial_observer = PositionRestrictedObservationWrapper([position_observer, team_observer, life_observer], **kwargs)
+
+        # Actor components
+        # These components handle the actions in the step function. This environment
+        # supports agents that can move around and attack agents from other teams.
+        self.move_actor = GridMovementActor(position=self.position_state, **kwargs)
+        self.attack_actor = AttackActor(**kwargs)
+
+        # Done components
+        # This component tracks when the simulation is done. This environment is
+        # done when either:
+        # (1) All the hunter have killed all the foragers.
+        # (2) All the foragers have killed all the resources.
+        # self.done = AnyTeamDeadDone(**kwargs)
+        self.done = TeamDeadDone(**kwargs)
+
+        # This is needed at the end of init in every environment. It ensures that
+        # agents have been configured correctly.
+        self.finalize()
+    
+    def reset(self, **kwargs):
[…]
