### 2021-07-09 01c912a
Updated hunter forager example
