### 2021-07-09 b1f4077
Managers track if the agents are observing and acting
