+++ setup.py
@@ -18,18 +18,18 @@ setuptools.setup(
         'Intended Audience :: Science/Research',
         'License :: OSI Approved :: BSD License',
         'Natural Language :: English',
-        'Operating System :: OS Independent',
+        'Operating System :: Unix',
         'Programming Language :: Python :: 3.7',
         'Topic :: Scientific/Engineering :: Artificial Intelligence',
     ],
     project_urls={
-        'Highlights': 'https://abmarl.readthedocs.io/en/latest/highlights.html',
+        'Featured Usage': 'https://abmarl.readthedocs.io/en/latest/featured_usage.html',
         'Documentation': 'https://abmarl.readthedocs.io/en/latest/index.html',
     },
     packages=setuptools.find_packages(),
     install_requires=[
-        'tensorflow==2.4.0',
-        'ray[rllib]==1.2.0',
+        'tensorflow',
+        'ray[rllib]==1.4.0',
         'matplotlib',
         'seaborn',
     ],
