### 2021-07-13 76d94a1
Updated hunting and foraging demo config
