### 2021-07-13 a1dcaa9
Added min harvest to harvesting agent
