### 2021-07-13 16c3c78
Upper bound obs space for GRidposition team based observer
