### 2021-07-13 0720b5b
Fixed ugly error in TurnBasedManager
