### 2021-07-13 fbd3470
Fixed ordering issue in TurnBasedManager agent cycle
