### 2021-07-13 fbbf9df
Fixed flake8 errors from hunting and foraging environment.
