+++ abmarl/sim/components/examples/hunting_and_foraging.py
@@ -4,13 +4,14 @@ import numpy as np
 
 # Import all the features that we need from the simulation components
 from abmarl.sim.components.state import GridPositionState, LifeState
-from abmarl.sim.components.observer import PositionObserver, LifeObserver, TeamObserver, GridPositionBasedObserver
+from abmarl.sim.components.observer import PositionObserver, LifeObserver, TeamObserver
 from abmarl.sim.components.wrappers.observer_wrapper import PositionRestrictedObservationWrapper
 from abmarl.sim.components.actor import GridMovementActor, AttackActor
-from abmarl.sim.components.done import AnyTeamDeadDone, TeamDeadDone
+from abmarl.sim.components.done import AnyTeamDeadDone
 
 # Environment needs a corresponding agent component
-from abmarl.sim.components.agent import ComponentAgent, AttackingAgent, GridMovementAgent, AgentObservingAgent, PositionObservingAgent, TeamObservingAgent, LifeObservingAgent
+from abmarl.sim.components.agent import ComponentAgent, AttackingAgent, GridMovementAgent, \
+    AgentObservingAgent, PositionObservingAgent, TeamObservingAgent, LifeObservingAgent
 
 # Import the interface
 from abmarl.sim import AgentBasedSimulation
@@ -18,16 +19,23 @@ from abmarl.sim import AgentBasedSimulation
 # Import extra tools
 from abmarl.tools.matplotlib_utils import mscatter
 
+
 # All HuntingForagingAgents
 # have a position, team, and life/death state
 # can observe positions, teams, and life state of other agents
 # can move around the grid and attack other agents
-class HuntingForagingAgent(AttackingAgent, GridMovementAgent, AgentObservingAgent, PositionObservingAgent, TeamObservingAgent, LifeObservingAgent): pass
+class HuntingForagingAgent(
+    AttackingAgent, GridMovementAgent, AgentObservingAgent, PositionObservingAgent,
+    TeamObservingAgent, LifeObservingAgent
+): pass
+
 
 # All FoodAgents
 # have a tema, position, and life
-# They are not really "agents" in the RL sense, they're just entities in the simulation for the foragers to gather.
+# They are not really "agents" in the RL sense, they're just entities in the
+# simulation for the foragers to gather.
 class FoodAgent(ComponentAgent): pass
+
 
 # Create the simulation environment from the components
 class HuntingForagingEnv(AgentBasedSimulation):
@@ -46,11 +54,12 @@ class HuntingForagingEnv(AgentBasedSimulation):
         # These components handle the observations that the agents receive whenever
         # get_obs is called. In this environment supports agents that can observe
         # the position, health, and team of other agents and itself.
-        self.partial_observer = GridPositionBasedObserver(position_state=self.position_state, **kwargs)
-        # position_observer = PositionObserver(position=self.position_state, **kwargs)
-        # team_observer = TeamObserver(**kwargs)
-        # life_observer = LifeObserver(**kwargs)
-        # self.partial_observer = PositionRestrictedObservationWrapper([position_observer, team_observer, life_observer], **kwargs)
+        position_observer = PositionObserver(position_state=self.position_state, **kwargs)
+        team_observer = TeamObserver(**kwargs)
+        life_observer = LifeObserver(**kwargs)
+        self.partial_observer = PositionRestrictedObservationWrapper(
+            [position_observer, team_observer, life_observer], **kwargs
+        )
 
         # Actor components
         # These components handle the actions in the step function. This environment
@@ -63,13 +72,12 @@ class HuntingForagingEnv(AgentBasedSimulation):
         # done when either:
[…]
