### 2021-07-13 fbf188d
Fixed linting problems
