+++ requirements.txt
@@ -1,4 +1,5 @@
 tensorflow
+gym
 ray[rllib]==1.4.0
 matplotlib
 seaborn
