### 2021-07-13 6791ac9
Removed gym from reqs and updated hunting foraging example demo
