+++ examples/hunting_foraging_demo.py
@@ -139,7 +139,7 @@ params = {
         'checkpoint_freq': 10,
         'checkpoint_at_end': True,
         'stop': {
-            'episodes_total': 200,
+            'episodes_total': 2000,
         },
         'verbose': 2,
         'config': {
@@ -152,7 +152,7 @@ params = {
                 'policies': policies,
                 'policy_mapping_fn': policy_mapping_fn,
             },
-            "num_workers": 0,
+            "num_workers": 7,
             "num_envs_per_worker": 1, # This must be 1 because we are not "threadsafe"
             "rollout_fragment_length": 200,
             "batch_mode": "complete_episodes",
