+++ requirements.txt
@@ -1,5 +1,4 @@
 tensorflow
-gym
 ray[rllib]==1.4.0
 matplotlib
 seaborn
