### 2021-07-19 75d4c9a
Added simple single-agent foraging example
