+++ abmarl/sim/examples/__init__.py
@@ -1,0 +1,1 @@
+from .foraging import ForagingSim
