+++ abmarl/sim/examples/foraging.py
@@ -1,0 +1,154 @@
+from enum import IntEnum
+
+from gym.spaces import MultiBinary, Box, Discrete
+from matplotlib import pyplot as plt
+import numpy as np
+
+from abmarl.sim import AgentBasedSimulation, Agent, PrincipleAgent
+from abmarl.tools.matplotlib_utils import mscatter
+
+class ForagingSim(AgentBasedSimulation):
+    class Actions(IntEnum):
+        UP = 0
+        RIGHT = 1
+        DOWN = 2
+        LEFT = 3
+        HARVEST = 4
+
+    def __init__(self):
+        self.region = 10
+        self.agents = {'forager': Agent(
+            id='forager',
+            action_space=Discrete(5),
+            observation_space={
+                'food': MultiBinary(5),
+                'position': Box(0, self.region, (2,), np.int)
+            }
+        )}
+
+        self.finalize()
+
+    def reset(self):
+        self.food = {}
+        for i in range(5):
+            self.food[f'food{i}'] = PrincipleAgent(
+                id=f'food{i}'
+            )
+        self.food['food0'].position = np.array([2, 2])
+        self.food['food1'].position = np.array([0, 6])
+        self.food['food2'].position = np.array([5, 4])
+        self.food['food3'].position = np.array([7, 3])
+        self.food['food4'].position = np.array([7, 7])
+        
+        self.agents['forager'].position = np.random.randint(0, self.region, 2)
+
+        self.reward = 0
+
+    def step(self, action):
+        forager = self.agents['forager']
+        if action == self.Actions.UP:
+            if forager.position[0] != 0:
+                forager.position += np.array([-1, 0])
+                self.reward = -1
+            else:
+                self.reward = -5
+        elif action == self.Actions.RIGHT:
+            if forager.position[1] != self.region - 1:
+                forager.position += np.array([0, 1])
+                self.reward = -1
+            else:
+                self.reward = -5
+        elif action == self.Actions.DOWN:
+            if forager.position[0] != self.region - 1:
+                forager.position += np.array([1, 0])
+                self.reward = -1
+            else:
+                self.reward = -5
+        elif action == self.Actions.LEFT:
+            if forager.position[1] != 0:
+                forager.position += np.array([0, -1])
+                self.reward = -1
+            else:
+                self.reward = -5
+        elif action == self.Actions.HARVEST:
+            if self._harvest_food(forager):
+                self.reward = 100
+            else:
+                self.reward = -5
+
+    def get_obs(self, agent_id):
+        agent = self.agents[agent_id]
+        food = []
+        for i in range(5):
+            if f'food{i}' in self.food:
+                food.append(1)
+            else:
+                food.append(0)
+        return {'food': food, 'position': agent.position}
+
+    def get_reward(self, agent_id):
+        reward = self.reward
+        self.reward = 0
+        return reward
+
+    def get_done(self, agent_id):
+        return not self.food
+
+    def get_all_done(self, **kwargs):
+        return not self.food
+
+    def get_info(self, agent_id, **kwargs):
+        return {}
+
+    def render(self, fig=None, **kwargs):
+        fig.clear()
+
+        ax = fig.gca()
+        ax.set(xlim=(0, self.region), ylim=(0, self.region))
+        ax.set_xticks(np.arange(0, self.region, 1))
+        ax.set_yticks(np.arange(0, self.region, 1))
+        ax.grid()
+
+        agents_x = [
+            agent.position[1] + 0.5 for agent in self.food.values()
+        ]
+        agents_y = [
+            self.region - 0.5 - agent.position[0] for agent in self.food.values()
+        ]
[…]
