### 2021-07-19 d645422
Added Q-Learning algorithm
