### 2021-07-19 e07a5db
Example 1 to fill in the blanks of the q-learning algorithm
