### 2021-07-19 LLNL/Abmarl#167 @abhiramm7
JOSS: Installation Instructions
Hi, I am following the installation instructions in the documentation and `pip install abmarl` is causing the following error,
```
ERROR: Cannot install abmarl==0.1.1 and abmarl==0.1.2 because these package versions have conflicting dependencies.

The conflict is caused by:
    abmarl 0.1.2 depends on ray[rllib]==1.2.0
    abmarl 0.1.1 depends on ray[rllib]==1.2.0

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/user_guide/#fixing-conflicting-dependencies
```

This is in OS X and python 3.9. I was able to install the package by removing `ray[rllib]==1.4.0` in the requirements. Is there a specific reason that Abmarl requires version 1.4.0? if not, consider replacing the == with >=. It also seems like there is a discrepancy between the dependency in the pypi.
