### 2021-07-19 #3424 @seba-1511
@rusu24edward Apologies for the delay -- so far the submission looks good. I think ray in requirements.txt should be updated to `ray[default]` (a warning otherwise pops up) but even with that, I get the following stack trace when following the quickstart instructions:

~~~shell
Traceback (most recent call last):
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/redis/connection.py", line 559, in connect
    sock = self._connect()
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/redis/connection.py", line 615, in _connect
    raise err
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/redis/connection.py", line 603, in _connect
    sock.connect(socket_address)
TimeoutError: [Errno 60] Operation timed out

Traceback (most recent call last):
  File "/Users/sebarnol/anaconda3/envs/abmarl36/bin/abmarl", line 33, in <module>
    sys.exit(load_entry_point('abmarl', 'console_scripts', 'abmarl')())
  File "/Users/sebarnol/Desktop/Abmarl/abmarl/scripts/scripts.py", line 43, in cli
    train.run(path_config)
  File "/Users/sebarnol/Desktop/Abmarl/abmarl/scripts/train_script.py", line 17, in run
    train.run(full_config_path)
  File "/Users/sebarnol/Desktop/Abmarl/abmarl/train.py", line 29, in run
    ray.init()
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/ray/_private/client_mode_hook.py", line 62, in wrapper
    return func(*args, **kwargs)
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/ray/worker.py", line 797, in init
    ray_params=ray_params)
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/ray/node.py", line 230, in __init__
    self.start_head_processes()
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/ray/node.py", line 861, in start_head_processes
    self.start_redis()
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/ray/node.py", line 686, in start_redis
    port_denylist=self._ray_params.reserved_ports)
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/ray/_private/services.py", line 891, in start_redis
    primary_redis_client.set("NumRedisShards", str(num_redis_shards))
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/redis/client.py", line 1801, in set
    return self.execute_command('SET', *pieces)
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/redis/client.py", line 898, in execute_command
    conn = self.connection or pool.get_connection(command_name, **options)
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/redis/connection.py", line 1192, in get_connection
    connection.connect()
  File "/Users/sebarnol/anaconda3/envs/abmarl36/lib/python3.6/site-packages/redis/connection.py", line 563, in connect
    raise ConnectionError(self._error_message(e))
redis.exceptions.ConnectionError: Error 60 connecting to 10.142.140.7:6379. Operation timed out.
[1]    87008 terminated  abmarl train examples/multi_corridor_example.py
~~~

Specifically, I ran the following:

~~~shell
conda create -n abmarl36 python=3.6
conda activate abmarl36
git clone https://github.com/LLNL/Abmarl/
cd Abmarl
pip install -r requirements.txt
pip install -e . --no-deps
abmarl train examples/multi_corridor_example.py
~~~
