### 2021-07-20 b8fcdf1
Added hint for creating q-table
