### 2021-07-20 #3424 @seba-1511
Other than that (and once I can confirm functionality), my only minor comment is on state of need/scholarship. The discussion around existing software around multi-agent RL could be more thorough; for example, here's some popular software that could be discussed:

* MADDPG, from OpenAI: https://github.com/openai/maddpg
* pymarl, from Oxford: https://github.com/oxwhirl/pymarl
* SMARTS, from Huawei: https://github.com/huawei-noah/SMARTS
* MAAC, from Shariq Iqbal: https://github.com/shariqiqbal2810/MAAC
* MarLo, from crowdAI: https://github.com/crowdAI/marLo
