### 2021-07-20 LLNL/Abmarl#168 @abhiramm7
JOSS: Functionality documentation
Hi, I commend the authors on thorough documentation. In the JOSS paper submission, you mention that Abmarl integration with open ai gym and MultiAgentEnv. Having some documentation or a simple example on how users can interface with the two packages might be helpful.
