### 2021-07-20 f61ca91
Deep Foraging Simulation for day 2 exercises
