+++ abmarl/sim/examples/deep_foraging.py
@@ -1,0 +1,233 @@
+
+from matplotlib import pyplot as plt
+import numpy as np
+
+# Import all the features that we need from the simulation components
+from abmarl.sim.components.state import GridPositionState, LifeState
+from abmarl.sim.components.observer import GridPositionBasedObserver
+from abmarl.sim.components.actor import GridMovementActor, AttackActor
+from abmarl.sim.components.done import AnyTeamDeadDone
+
+# Environment needs a corresponding agent component
+from abmarl.sim.components.agent import ComponentAgent, AttackingAgent, GridMovementAgent, \
+    AgentObservingAgent, PositionObservingAgent, TeamObservingAgent, LifeObservingAgent
+
+# Import the interface
+from abmarl.sim import AgentBasedSimulation
+
+# Import extra tools
+from abmarl.tools.matplotlib_utils import mscatter
+
+
+# All HuntingForagingAgents
+# have a position, team, and life/death state
+# can observe positions, teams, and life state of other agents
+# can move around the grid and attack other agents
+class Forager(
+    AttackingAgent, GridMovementAgent, AgentObservingAgent, PositionObservingAgent,
+    TeamObservingAgent, LifeObservingAgent
+):
+    def __init__(self, **kwargs):
+        kwargs['team'] = 2
+        kwargs['move_range'] = 1
+        kwargs['attack_strength'] = 1
+        super().__init__(**kwargs)
+
+
+# All FoodAgents
+# have a tema, position, and life
+# They are not really "agents" in the RL sense, they're just entities in the
+# simulation for the foragers to gather.
+class FoodAgent(ComponentAgent):
+    def __init__(self, **kwargs):
+        kwargs['team'] = 1
+        super().__init__(**kwargs)
+
+
+# Create the simulation environment from the components
+class DeepForagingSim(AgentBasedSimulation):
+    def __init__(self, num_food, **kwargs):
+        # Hack the number of food agents
+        food = {f'food{i}': FoodAgent(id=f'food{i}') for i in range(num_food)}
+        kwargs['agents'] = {**kwargs['agents'], **food}
+        # Explicitly pull out the the dictionary of agents. This makes the env
+        # easier to work with.
+        self.agents = kwargs['agents']
+
+        # Hack the team attack matrix and number of teams
+        team_attack_matrix = np.zeros((3, 3))
+        team_attack_matrix[2, 1] = 1 # Foragers can attack food
+        kwargs['team_attack_matrix'] = team_attack_matrix
+        kwargs['number_of_teams'] = 2
+
+        # State components
+        # These components track the state of the agents. This environment supports
+        # agents with positions, life, and team.
+        self.position_state = GridPositionState(**kwargs)
+        self.life_state = LifeState(**kwargs)
+
+        # Observer components
+        # These components handle the observations that the agents receive whenever
+        # get_obs is called. In this environment supports agents that can observe
+        # the position, health, and team of other agents and itself.
+        self.observer = GridPositionBasedObserver(position_state=self.position_state, **kwargs)
+
+        # Actor components
+        # These components handle the actions in the step function. This environment
+        # supports agents that can move around and attack agents from other teams.
+        self.move_actor = GridMovementActor(position_state=self.position_state, **kwargs)
+        self.attack_actor = AttackActor(**kwargs)
+
+        # Done components
+        # This component tracks when the simulation is done. This environment is
+        # done when either:
+        # (1) All the hunter have killed all the foragers.
+        # (2) All the foragers have killed all the resources.
+        self.done = AnyTeamDeadDone(**kwargs)
+
[…]
