+++ abmarl/sim/examples/__init__.py
@@ -1,1 +1,2 @@
 from .foraging import ForagingSim
+from .deep_foraging import DeepForagingSim, Forager
