### 2021-07-20 9746d1d
Exercise 3 examining the features of deep foraging sim
