### 2021-07-20 298bf0e
Enhanced the gym wrapper to assert only a single LEARNING agent
