### 2021-07-20 4d0db7a
Training config for deep forager
