### 2021-07-20 775cd16
Deep foraging uses discrete action for forager
