### 2021-07-20 d32be74
Added exercise 4 position analysis
