### 2021-07-21 #3424 @rusu24edward
Thanks for the reviews! I will try to address each of these points by the end of the week.
