### 2021-07-21 #3424 @rusu24edward
@seba-1511 Regarding your comment on other popular software that can be addressed, I'd be happy to include more discussion on comparisons with popular software, but I think it's important to clarify the suggestions you made as it will shed light on what ABMARL targets. Reinforcement learning has two components: simulation and algorithm (also commonly referred to as environment and agent). Our software targets the simulation component, so I restricted my discussion to other software that specifically targets the simulation component.

Regarding your suggestions:

**MADDPG**
This paper actually has both simulation and algorithm. The authors create a particle simulation (MPE) and a multi-agent algorithm that they benchmark against their simulation. While this is a popular paper, it is no longer maintained by the originators. The algorithm has since been folded in to RLlib, to which ABMARL connects, and the simulation has been folded into the Petting Zoo, which we do directly mention in our paper.
MADDPG actually highlights the importance of ABMARL. The creators of MADDPG chose to create a simulation in addition to the algorithm in order to benchmark their work. They made it from from scratch, and I argue that the simulation interface that Abmarl provides will greatly ease the burden on algorithm developers who want simple test beds for their multi-agent algorithms. (maybe this is a good sentence to add to the paper?)

**pymarl**
PyMarl is a collection of multi-agent algorithms, and each of the algorithms that they support is also supported in RLlib, to which ABMARL connects. Pymarl does not have its own simulation benchmark suite; it uses StarCraft, which we reference in our paper.

**MAAC**
Similar to pymarl, MAAC targets the algorithm component of RL, and the algorithm is supported in RLlib, to which ABMARL connects. It does not really create its own simulation suite. Instead, it creates a specific scenario within MPE (which is the simulation from MADDPG), which is folded into PettingZoo, which we do reference in our paper.

**SMARTS** and **MARLO**
Both of these packages do indeed target the simulation component of RL, and I would be happy to include them in my comparison if you think that would be beneficial. Like MAgent, Starcraft, and Neural MMO (already addressed in our paper), both of these also couple the interface with a simulation. I can add these two to the list that I mention in paragraph 2 of statement of need, or I can also expound on each of the packages I listed (include smarts and marlo). I tried to keep it short and sweet in the flavor of many JOSS papers I see, but I'm happy to add more. Suggestions?
