### 2021-07-21 LLNL/Abmarl#167 @rusu24edward
Hi @abhiramm7,

Thanks for bringing this to my attention. It sounds like you tried to install abmarl from pip after having already installed ray[rllib] 1.4.0. I just tested the installation in a clean virtual environment, and I did not have any error. Here is what I did:
```
python3 -m venv my_venv
source my_venv/bin/activate
pip install --upgrade pip # Make sure you have newest version of pip
pip install abmarl
```
No conflict there.

To answer your question about the versioning: The released version (branch Abmarl-87-interface-release) uses Ray 1.2.0, but our main branch (which is our development branch) is relying on version 1.4.0. This is because there is new content in our main branch that relies on features in ray 1.4.0 that did not exist in 1.2.0. That is why there is a discrepancy. The version of abmarl on pypi comes from the release branch 87.
