### 2021-07-21 #3424 @rusu24edward
> Specifically, I ran the following:
>
> ```shell
> conda create -n abmarl36 python=3.6
> conda activate abmarl36
> git clone https://github.com/LLNL/Abmarl/
> cd Abmarl
> pip install -r requirements.txt
> pip install -e . --no-deps
> abmarl train examples/multi_corridor_example.py
> ```

@seba-1511 Thanks for bringing this to my attention. I just ran a fresh virtual environment and did not get any error. We technically specify that python 3.7+ is required, can you try this again with python3.7?
