### 2021-07-21 LLNL/Abmarl#168 @rusu24edward
HI @abhiramm7

Thank you for your insightful review. Regarding integration with `MultiAgentEnv` and `gym.Env`, we do have this documented in the API spec [here](https://abmarl.readthedocs.io/en/latest/api.html#abmarl.external.MultiAgentWrapper) and [here](https://abmarl.readthedocs.io/en/latest/api.html#abmarl.external.GymWrapper). Additionally, we the [MultiCorridor tutorial](https://abmarl.readthedocs.io/en/latest/tutorials/multi_corridor.html#simulation-setup) and [PredatorPrey tutorial](https://abmarl.readthedocs.io/en/latest/tutorials/predator_prey.html#experiment-parameters) show how to use the `MultiAgentWrapper` to treat the simulation as a `MultiAgentEnv`.

Does the documentation need to be clarified? Should I add another example using the GymWrapper?
