### 2021-07-21 0b3c78a
Added links to the files for the tutorials
