### 2021-07-21 1d4cc72
Added hunting and foarging sim with exercise 5
