+++ abmarl/sim/examples/forager_hunter.py
@@ -1,0 +1,224 @@
+
+from enum import IntEnum
+
+from gym.spaces import Discrete
+from matplotlib import pyplot as plt
+import numpy as np
+
+# Import all the features that we need from the simulation components
+from abmarl.sim.components.state import GridPositionState, LifeState
+from abmarl.sim.components.observer import GridPositionBasedObserver
+from abmarl.sim.components.actor import GridMovementActor, AttackActor
+from abmarl.sim.components.done import AnyTeamDeadDone
+
+# Environment needs a corresponding agent component
+from abmarl.sim.components.agent import ComponentAgent, AttackingAgent, GridMovementAgent, \
+    AgentObservingAgent, PositionObservingAgent, TeamObservingAgent, LifeObservingAgent
+
+# Import the interface
+from abmarl.sim import AgentBasedSimulation
+
+# Import extra tools
+from abmarl.tools.matplotlib_utils import mscatter
+
+
+# All HuntingForagingAgents
+# have a position, team, and life/death state
+# can observe positions, teams, and life state of other agents
+# can move around the grid and attack other agents
+class Forager(
+    AttackingAgent, GridMovementAgent, AgentObservingAgent, PositionObservingAgent,
+    TeamObservingAgent, LifeObservingAgent
+):
+    def __init__(self, **kwargs):
+        kwargs['team'] = 2
+        kwargs['move_range'] = 1
+        kwargs['attack_strength'] = 1
+        super().__init__(**kwargs)
+
+class Hunter(
+    AttackingAgent, GridMovementAgent, AgentObservingAgent, PositionObservingAgent,
+    TeamObservingAgent, LifeObservingAgent
+):
+    def __init__(self, **kwargs):
+        kwargs['team'] = 3
+        kwargs['move_range'] = 1
+        kwargs['attack_strength'] = 1
+        super().__init__(**kwargs)
+
+
+# All FoodAgents
+# have a tema, position, and life
+# They are not really "agents" in the RL sense, they're just entities in the
+# simulation for the foragers to gather.
+class FoodAgent(ComponentAgent):
+    def __init__(self, **kwargs):
+        kwargs['team'] = 1
+        super().__init__(**kwargs)
+
+
+# Create the simulation environment from the components
+class HuntingForagingSim(AgentBasedSimulation):
+    class Actions(IntEnum):
+        UP = 0
+        RIGHT = 1
+        DOWN = 2
+        LEFT = 3
+        ATTACK = 4
+
+    _action_mapping = {
+        Actions.UP: {'move': np.array([-1, 0]), 'attack': False},
+        Actions.RIGHT: {'move': np.array([0, 1]), 'attack': False},
+        Actions.DOWN: {'move': np.array([1, 0]), 'attack': False},
+        Actions.LEFT: {'move': np.array([0, -1]), 'attack': False},
+        Actions.ATTACK: {'move': np.array([0, 0]), 'attack': True}
+    }
+    
+    def __init__(self, num_food, **kwargs):
+        # Hack the number of food agents
+        food = {f'food{i}': FoodAgent(id=f'food{i}') for i in range(num_food)}
+        kwargs['agents'] = {**kwargs['agents'], **food}
+        # Explicitly pull out the the dictionary of agents. This makes the env
+        # easier to work with.
+        self.agents = kwargs['agents']
+
+        # Hack the team attack matrix and number of teams
+        team_attack_matrix = np.zeros((4, 4))
+        team_attack_matrix[2, 1] = 1 # Foragers can attack food
+        team_attack_matrix[3, 2] = 1 # Hunters can attack foragers
+        kwargs['team_attack_matrix'] = team_attack_matrix
+        kwargs['number_of_teams'] = 3
+
+        # State components
[…]
