### 2021-07-21 7ae48a1
Added installation instructions for the class
