### 2021-07-21 0f9ba96
Added exercise 6
