### 2021-07-22 3febe13
LC users in docs for setup
