### 2021-07-22 daf3674
Better split in setup docs
