+++ docs/src/dssi_rl_setup.rst
@@ -5,57 +5,15 @@
 DSSI RL Setup
 =============
 
-Unix Users (LC Users see below)
--------------------------------
+Registered Attendees
+--------------------
 
-If you have Mac or Linux, you can install ABMARL directly onto your machine.
-
-First, download the repository:
-
-.. code-block::
-
-   git clone https://github.com/LLNL/Abmarl.git
-   cd Abmarl
-   git checkout abmarl-dssi-class
-
-
-Then, create a virtual environment and install ABMARL from the source
-
-.. code-block::
-
-   python3 -m venv v_abmarl
-   source v_abmarl/bin/activate
-   pip3 install --upgrade pip3
-   pip3 install -r requirements.txt
-   pip3 install -e . --no-deps
-
-Test that your installation is successful
-
-.. code-block::
-
-   abmarl train examples/multi_corridor_example.py
-
-This will create `abmarl_results` in your home directory and will store the trained
-behavior in a subdirectory with the name of the simulation, `MultiCorridor`, followed
-by the date and time (e.g. `MultiCorridor_2021-07-20_17-36`). You can confirm the
-trained behavior with
-
-.. code-block::
-
-   abmarl visualize ~/abmarl_results/MultiCorridor_2021-07-20_17-36/
-
-You should see five green squares moving to the right.
-
-
-Windows/Amazon Workspace Users/LC Users
----------------------------------------
-
-If you have Windows and/or are using the Amazon Workspace, you can connect to CZ
-via VNC and use the ABMARL installation on CZ. You must have an LC account and
+If you are a DSSI student or have contacted me before the class, then you will have access
+to the installation I built on the CZ. You must have an LC account and
 `VNC viewer <https://hpc.llnl.gov/software/visualization-software/vnc-realvnc>`_
 installed on your system.
 
-Once you've done that, ssh onto the system of choice (probably flash for
+Once you've done that, ssh onto the CZ system of choice (probably flash for
 DSSI students).
 
 .. code-block
@@ -92,3 +50,52 @@ trained behavior with
    abmarl visualize ~/abmarl_results/MultiCorridor_2021-07-20_17-36/
 
 You should see five green squares moving to the right.
+
+
+Additional Attendees
+--------------------
+
+If you have Mac or Linux, you can install ABMARL directly onto your machine. Otherwise,
+you will need to install it on a CZ machine.
+
+First, download the repository:
+
+.. code-block::
+
+   git clone https://github.com/LLNL/Abmarl.git
+   cd Abmarl
+   git checkout abmarl-dssi-class
+
+
+Then, create a virtual environment and install ABMARL from the source
+
+.. ATTENTION:
+   If you are installing on LC, you should setup your virtual environment using
+   the system site packages, like so ``python3 -m venv --system-site-packages v_abmarl``
+
+.. code-block::
+
+   python3 -m venv v_abmarl
+   source v_abmarl/bin/activate
+   pip3 install --upgrade pip3
+   pip3 install -r requirements.txt
+   pip3 install -e . --no-deps
+
+Test that your installation is successful
+
+.. code-block::
+
+   abmarl train examples/multi_corridor_example.py
+
+This will create `abmarl_results` in your home directory and will store the trained
+behavior in a subdirectory with the name of the simulation, `MultiCorridor`, followed
+by the date and time (e.g. `MultiCorridor_2021-07-20_17-36`). You can confirm the
+trained behavior with
+
+.. code-block::
+
+   abmarl visualize ~/abmarl_results/MultiCorridor_2021-07-20_17-36/
+
+You should see five green squares moving to the right.
+
+
