### 2021-07-26 #3424 @rusu24edward
@abhiramm7 Thank you for your review so far. I believe I have followed up on all the points you have mentioned above. Please see my responses to the issues you created in Abmarl [here](https://github.com/LLNL/Abmarl/issues/167) and [here](https://github.com/LLNL/Abmarl/issues/168). Please let me know if there is still more to be done on these points or on any other part of the submission.

@seba-1511 Thank you for your review so far. My follow up was primarily asking for clarification [here](https://github.com/openjournals/joss-reviews/issues/3424#issuecomment-883812828) and requesting that you try the build again with Python3.7 [here](https://github.com/openjournals/joss-reviews/issues/3424#issuecomment-883880317). Please take a look when you get a chance.
