### 2021-07-26 78abc9e
Both grid types use np array of dictionaries. Distinction based on logic
