### 2021-07-26 62beafe
Single grid class with single logic. Overlappability is determined by the agents themselves.
