### 2021-07-26 5287a39
Single grid class with single logic. Overlappability is determined by the agents themselves.
