+++ abmarl/sim/gridworld/grid.py
@@ -1,26 +1,34 @@
 
-from abc import ABC, abstractmethod
-
 import numpy as np
 
-class Grid(np.ndarray, ABC):
+class Grid:
     """
-    A Grid is a numpy array that stores the positions of the agents.
+    A Grid has a numpy array that stores the agents at indicies of the array.
 
-    Manages the agent positions through the add and remove functions while abstracting
-    the underlying datatype used to store the agent there.
+    Components can interface with the internals of the Grid. Each index in the
+    grid is a dictionary that maps the agent id to the agent object itself. If agents
+    can overlap, then there may be more than one agent per cell.
+
+    Args:
+        rows: The number of rows in the grid.
+        cols: The number of columns in the grid.
     """
-    @abstractmethod
+    def __init__(self, rows, cols):
+        self._internal = np.empty((rows, cols), dtype=object)
+
     def reset(self, **kwargs):
         """
         Reset the grid to an empty state.
         """
-        pass
+        for i in range(self.shape[0]):
+            for j in range(self.shape[1]):
+                self[i,j] = {}
 
-    @abstractmethod
     def query(self, agent, ndx):
         """
-        Query a cell in the grid to see if is available to this agent.
+        Query a cell in the grid to see if is available to this agent. The cell
+        is available for the agent if it is empty or if both the occupying agent
+        and the querying agent are overlappable.
 
         Args:
             agent: The agent for which we are checking availabilty.
@@ -29,15 +37,19 @@ class Grid(np.ndarray, ABC):
         Returns:
             True if the cell is available for this agent, otherwise False.
         """
-        pass
+        ndx = tuple(ndx)
+        return not self[ndx] or (
+            next(iter(self[ndx].values())).overlappable and agent.overlappable
+        )
 
-    @abstractmethod
     def place(self, agent, ndx):
         """
         Place an agent at an index.
 
         If the placement is successful, the agent will be placed at that index
-        in the grid and the agent's position will be updated.
+        in the grid and the agent's position will be updated. The placement is
+        successful if the new position is unoccupied or if the agent already occupying
+        that position is overlappable AND this agent is overlappable.
 
         Args:
             agent: The agent to place.
@@ -46,15 +58,22 @@ class Grid(np.ndarray, ABC):
         Returns:
             The successfulness of the placement.
         """
-        pass
+        ndx = tuple(ndx)
+        if self.query(agent, ndx):
+            self._place(agent, ndx)
+            agent.position = np.array(ndx)
+            return True
+        else:
+            return False
 
-    @abstractmethod
     def move(self, agent, to_ndx):
         """
         Move an agent from one index to another.
 
         If the move is successful, the agent will be placed at that index in the
-        grid and the agent's position will be updated.
+        grid and the agent's position will be updated. The move is successful if
+        the new position is unoccupied or if the agent already occupying that position
+        is overlappable AND this agent is overlappable.
 
         Note: the agent will be removed from its old position in the grid.
 
@@ -64,118 +83,6 @@ class Grid(np.ndarray, ABC):
 
         Returns:
             The successfulness of the move.
-        """
-        pass
-
-    @abstractmethod
-    def remove(self, agent, ndx):
-        """
-        Remove an agent from an index.
-
-        Args:
-            agent: The agent to remove
-            ndx: The old index for this agent.
-        """
-        pass
-
-    @abstractmethod
-    def _place(self, agent, ndx):
-        """
-        Unprotected placement. Internal use only.
-        """
-        pass
-
-
-class NonOverlappingGrid(Grid):
-    """
-    A grid where agents cannot overlap.
[…]
