### 2021-07-26 d775e40
Attack actor logic uses grid internal
