### 2021-07-27 LLNL/Abmarl#168 @abhiramm7
I see that the MultiCorridor uses gym.spaces. I was under the impression that abmarl would be able to wrap around an existing open ai gym environment. I apologize if I am mis-interpreting it. I was thinking an example that uses an existing open ai gym environment with abmarl would be helpful to users.
