### 2021-07-27 LLNL/Abmarl#168 @rusu24edward
Thanks for the clarification! I agree, this would be helpful. I'll add a page on running Abmarl with already-existing gym environments.
