### 2021-07-27 LLNL/Abmarl#167 @rusu24edward
Python 3.9 does not have wheels for ray 1.2.0 or tensorflow 2.4.0. When we try to install Abmarl, pip will try the latest  version, which will fail, and then try the previous version, which will also fail. Pip only has python 3.9 wheels for tensorflow 2.5.0+ and ray 1.4.1+. The release was built against ray 1.2.0 and tensorflow 2.4.0. We ensure that our releases will work against the dependencies we test them on, and so I believe the best solution is to specify Python 3.7 in Abmarl's setup for the initial release and work to get it running on newer versions for future releases (we already have the dev branch working on ray 1.4.0)
