### 2021-07-27 6ccfc6f
Version 0.1.3 corrects dependencies in setup
