### 2021-07-27 LLNL/Abmarl#167 @rusu24edward
@abhiramm7 because of breaks in newer versions of our dependencies (namely ray), we are restricted from upgrading abmarl to the newest version of ray dependencies, which keeps us from running on python 3.9 (see note above). I've added better version specification to the setup and documentation--we support python 3.7 and 3.8. I've tested this and released Abmarl 0.1.3 to Pypi, which should have the correct dependencies. Now, if you try to install abmarl with python 3.9, you will get an error. Installations using python versions 3.7 and 3.8 should work just fine.
