### 2021-07-27 03d2be8
Added an example to run gym's guessing game
