### 2021-07-28 59634c5
Fixed deep forager exercise 3
