### 2021-07-28 8314c04
train uses local dir as base directory for output
