### 2021-07-29 #3424 @seba-1511
@rusu24edward Thanks for the update.

Re related works: Thanks for the detailed answer. I think it's fine to keep the paper mostly as-is but since some of the above works also implement RL tasks I think they are worth mentioning, if only to highlight the need for Abmarl's simulation interface. This could be just 2-3 sentences, no need for a detailed comparison as above.

Re installation issues: Changing to Python 3.7 and upgrading Abmarl to the latest version (0.1.3) seems to have fixed it. Thank you.
