+++ publications/paper.md
@@ -39,26 +39,45 @@ In 2016, @gym published OpenAi Gym, an interface for single-agent simulations. T
 quickly became one of the most popular connections between simulation and training
 in RL experimentation. It has been used by many simulation benchmarks
 for single-agent reinforcement learning, including the Arcade Learning Environment [@arcade].
-Since then the field of Deep Reinforcemenet Learning (DRL) has exploded in both
+Since then the field of Deep Reinforcement Learning (DRL) has exploded in both
 algorithm development and simulation design, and over the past few years researchers
 have been extending their interest to Multi-Agent Reinforcement Learning (MARL).
-
-MARL has shown exceptional promise towards artificial
-general intelligence. Surprisingly complex and hierarchical behavior emerges in the
+Surprisingly complex and hierarchical behavior emerges in the
 interaction among multiple agents, especially when those agents differ in their
-objectives [@hide-n-seek]. Several projects have attempted to define a standard set
-of benchmark scenarios for Multi-Agent problems; such as MAgent [@magent], Starcraft [@smac], and
-Neural MMO [@neuralmmo]. However, each of these couples the interface with the 
-underlying simulation. Notably, @pettingzoo have attempted to unify some of
+objectives [@hide-n-seek]. 
+
+To train agents with Multi-Agent Reinforcement Learning, one needs two components: simulation
+and algorithm (also commonly referred to as environment and agent). Much effort
+has been given to the development of RL algorithms, which has brought us exciting
+breakthroughs and enhancements in the field of artificial intelligence. Our aim,
+however, focuses on the simulation component of Reinforcement Learning.
+
+Several projects have attempted to define a standard set
+of benchmark scenarios for Multi-Agent problems. In their groundbreaking work,
+@maddpg introduced MADDPG, a "centralized training, decentralized execution" multi-agent
+algorithm. Along with their algorithmic development, they created Multi-Particle
+Environment (MPE) (now managed as a part of PettingZoo) as a benchmark suite that
+includes continuous movement and communication features. @magent produced scalable
+grid-based simulations and demonstrated emergent behavior in multi-team games
+on the order of millions of agents. @smac and @marlo brought RL research closer to home,
+giving researchers access to dozens of multi-agent scenarios in the popular game
+StarCraft and Minecraft, respectively. @neuralmmo also targeted video games by
+supporting multi-agent reinforcement learning in MMORPG-styled simulations with
+persistent, open-ended tasks among multiple agents. @smarts brought us realistic
+traffic simulation scenarios to lead society towards autonomous driving.
+
+Each of these efforts and more are great milestones in RL simulation. Naturally,
+each of them couples the simulation interface with the 
+underlying simulation. @pettingzoo have attempted to unify some of
 the more popular simulations under a single interface, giving researchers easier
 access to these simulations. While this is a step towards
-a standard multi-agent interace, these are still tied to a specific set of already-built
-simulations with limited flexibility.
+a standard multi-agent interface, most simulation efforts are tied to a specific
+set of already-built simulations with limited flexibility.
 
 Abmarl defines an interface for multi-agent simulations that is versatile,
-extendible, and intuitive. Rather than adapting gym's interface for a targetted
+extendible, and intuitive. Rather than adapting gym's interface for a targeted
 multi-agent simulation, we have built an interface from scratch that allows for
-the greatest flexbility while connecting to one of the most advanced, general-purpose, and
+the greatest flexibility while connecting to one of the most advanced, general-purpose, and
 open-source libraries: RLlib [@rllib]. Our interface manages the loop between agents
 and the trainer, enabling the researcher to focus on simulation design and algorithmic
 development without worrying about the data exchange.
[…]
