### 2021-07-30 1d729b0
Small fixes in the paper
