### 2021-07-30 #3424 @rusu24edward
@seba-1511 Thank you very much for your insights on the paper. I have updated it, and it now includes a short explanation on the two components of RL and a more detailed examination of related work. Take a look and let me know what you think. I see that you have checked everything on the checklist; please let me know if there's anything else I can do.

@abhiramm7 I hope my recent changes to the dependencies and documentation help clarify the environment in which Abmarl is supported. I think there are still a few things in your checklist, please take a look when you can and let me know if there's anything else I can do.

Thank you both!
