### 2021-08-06 088d8b1
updated manager init file to import DynamicOrder objects
