+++ abmarl/managers/__init__.py
@@ -1,3 +1,5 @@
 from .simulation_manager import SimulationManager
 from .turn_based_manager import TurnBasedManager
 from .all_step_manager import AllStepManager
+from .dynamic_order_manager import DynamicOrderManager, DynamicOrderSimulation
+
