### 2021-08-11 #3424 @drvinceknight
Thanks, the paper itself looks good but there are a number of missing DOIs that would be great if you could fix please.

For example `OpenGym` has the following as their citation instruction: https://github.com/openai/gym#citation

```
@misc{1606.01540,
  Author = {Greg Brockman and Vicki Cheung and Ludwig Pettersson and Jonas Schneider and John Schulman and Jie Tang and Wojciech Zaremba},
  Title = {OpenAI Gym},
  Year = {2016},
  Eprint = {arXiv:1606.01540},
}
```

Although `arXiv:1606.01540` is not technically a DOI it should still appear in the rendered pdf please.

There are a number of other references where the url to the paper is given but not the DOI, we'll need to change those please (or let me know if you're unable to find the DOI).
