+++ publications/paper.bib
@@ -3,6 +3,8 @@
   Title = {OpenAI Gym},
   Year = {2016},
   Eprint = {arXiv:1606.01540},
+  url={https://arxiv.org/abs/1606.01540},
+  pdf={https://arxiv.org/pdf/1606.01540.pdf},
 }
 
 @InProceedings{rllib,
@@ -34,7 +36,7 @@
 @misc{hide-n-seek,
   title = {Emergent Tool Use From Multi-Agent Autocurricula},
   author = {Bowen Baker, Ingmar Kanitscheider, Todor Markov, Yi Wu, Glenn Powell, Bob McGrew, Igor Mordatch},
-  url = {https://arxiv.org/abs/1909.07528},
+  url = {https://openreview.net/pdf?id=SkxpxJBKwS},
   year = {2020},
   mont = {February}
 }
@@ -56,6 +58,8 @@
   journal = {Computing Research Repository},
   volume = {abs/1902.04043},
   year = {2019},
+  url = {https://arxiv.org/abs/1902.04043},
+  pdf = {https://arxiv.org/pdf/1902.04043.pdf},
 }
 
 @misc{neuralmmo,
@@ -69,7 +73,8 @@
   Title = {PettingZoo: Gym for Multi-Agent Reinforcement Learning},
   Author = {Terry, Justin K and Black, Benjamin and Jayakumar, Mario and Hari, Ananth and Santos, Luis and Dieffendahl, Clemens and Williams, Niall L and Lokesh, Yashas and Sullivan, Ryan and Horsch, Caroline and Ravi, Praveen},
   journal={arXiv preprint arXiv:2009.14471},
-  year={2020}
+  year={2020},
+  url={https://arxiv.org/pdf/2009.14471.pdf}
 }
 
 @misc{hybrid,
@@ -84,7 +89,7 @@
   author       = {Ryan Lowe and Yi Wu and Aviv Tamar and Jean Harb and Pieter Abbeel and Igor Mordatch},
   title        = {Multi-Agent Actor-Critic for Mixed Cooperative-Competitve Environments},
   year         = {2020},
-  url          = {https://arxiv.org/pdf/1706.02275.pdf}
+  url          = {https://papers.nips.cc/paper/2017/file/68a9750337a418a86fe06c1991a1d64c-Paper.pdf}
 }
 
 @misc{smarts,
