### 2021-08-11 #3424 @rusu24edward
Hi @drvinceknight, I updated the references so that each of them has a URL, only using an arxiv url when I was not able to find a different one. I was not able to find DOI's for any other papers. I used google's search and [cross ref](https://www.crossref.org/guestquery/), and only the arcade learning environment shows a DOI. Many of the references are conference proceedings, which I'm not sure if those get DOI's.
