### 2021-08-13 #3424 @rusu24edward
@drvinceknight I have released version 0.1.4, and here is the DOI: 10.5281/zenodo.5196791
