### 2021-08-14 b25c385
Cleaned place function from Grid
