### 2021-08-14 eb9d5c5
Cleaned remove function in grid
