### 2021-08-15 a370b01
Removed move command from grid and brought back remove and modified place
