### 2021-08-15 697ec4a
Grid subsets output the grid itself
