+++ abmarl/sim/gridworld/grid.py
@@ -17,6 +17,15 @@ class Grid:
     def __init__(self, rows, cols):
         self._internal = np.empty((rows, cols), dtype=object)
 
+    def __init__(self, array):
+        """
+        Create a grid from an already-created numpy array.
+
+        Args:
+            array: Numpy array to be used as the internal.
+        """
+        self._internal = array
+
     def reset(self, **kwargs):
         """
         Reset the grid to an empty state.
@@ -77,3 +86,16 @@ class Grid:
         """
         ndx = tuple(ndx)
         del self._internal[ndx][agent.id]
+
+    def __getitem__(self, subscript):
+        array = self._internal.__getitem__(subscript)
+        if type(array) is not np.ndarray: # It's a number retrieved from a single index
+            return array # Just return that number
+        else:
+            return Grid(array)
+
+    def __setitem__(self, subscript, item):
+        raise RuntimeError("You should not be setting an item directly in the grid.")
+
+    def __delitem__(self, subscript):
+        raise RuntimeError("You should not be deleting an item directly in the grid.")
