### 2021-08-16 3439ff5
Removed non-overlapping state because the overlappable state has logic that handles both cases.
