### 2021-08-16 5ca33fd
Attack Actor and Grid Observer use local grid correctly
