### 2021-08-16 a9e753d
Agent requires encoding and added support for render color and shape
