### 2021-08-16 f7f4c81
Removed overlapping arg from build sim since it's no longer needed
