### 2021-08-16 6312a2c
Bunch of little fixes to initialize and reset the grid.
