+++ abmarl/sim/gridworld/observer.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from abmarl.sim.gridworld.base import GridWorldBaseComponent
 from abmarl.sim.gridworld.agent import GridObservingAgent
-from abmarl.sim.gridworld.state import UniquePositionState
+from abmarl.sim.gridworld.state import PositionState
 
 
 class ObserverBaseComponent(GridWorldBaseComponent, ABC):
