### 2021-08-17 5ea08de
Lots of little changes from the PR review
