+++ abmarl/sim/gridworld/base.py
@@ -1,5 +1,5 @@
 
-from abc import ABC, abstractmethod
+from abc import ABC
 
 import numpy as np
 
