+++ abmarl/sim/gridworld/examples/build_grid_example.py
@@ -4,8 +4,8 @@ from matplotlib import pyplot as plt
 from abmarl.sim.gridworld.examples.gridworld_example import WallAgent, FoodAgent, GridSim
 
 object_registry = {
-    'A': lambda n: FoodAgent(id=f'explorer{n}', initial_health=1),
-    'W': lambda n: WallAgent(id=f'wall{n}')
+    'A': lambda n: FoodAgent(id=f'explorer{n}', initial_health=1, encoding=1),
+    'W': lambda n: WallAgent(id=f'wall{n}', encoding=1, render_shape='X')
 }
 
 file_name = 'starting_grid.txt'
