+++ abmarl/sim/gridworld/observer.py
@@ -55,13 +55,210 @@ class ObserverBaseComponent(GridWorldBaseComponent, ABC):
 class SingleGridObserver(ObserverBaseComponent):
     """
     Observe a subset of the grid centered on the agent's position.
+
+    The observation is centered around the observing agent's position. Each agent
+    in the "observation window" is recorded in the relative cell using its encoding.
+    For example, an agent might observe something like:
+        -1 -1 -1 -1 -1
+         0  2  0  0 -1
+        -2  1  3  0 -1
+         0  0  0  0 -1
+         0  2  2  0 -1
+    
+    If there are multiple agents on a single cell with different encodings, the
+    agent will observe only one of them chosen at random.
     """
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
         for agent in self.agents.values():
             if isinstance(agent, self.supported_agent_type):
                 agent.observation_space[self.key] = Box(
-                    -np.inf, np.inf, (agent.view_range, agent.view_range), np.int
+                    -np.inf, np.inf, (agent.view_range * 2 + 1, agent.view_range * 2 + 1), np.int
+                )
+
+    @property
+    def key(self):
+        """
+        This Observer's key is "grid".
+        """
+        return 'grid'
+
+    @property
+    def supported_agent_type(self):
+        """
+        This Observer works with GridObservingAgents.
+        """
+        return GridObservingAgent
+
+    def get_obs(self, agent, **kwargs):
+        """
+        The agent observes a sub-grid centered on its position.
+
+        The observation may include other agents, empty spaces, out of bounds, and
+        masked cells, which can be blocked from view by other view-blocking agents.
+
+        Returns:
+            The observation as a dictionary.
+        """
+        if not isinstance(agent, self.supported_agent_type):
+            return {}
+
+        # Generate a completely empty grid
+        # Fill the grid with out-of-bounds values, which will then be replaced by
+        # objects and empty space below.
+        local_grid = np.empty((agent.view_range * 2 + 1, agent.view_range * 2 + 1), dtype=object)
+
+        # Copy the section of the grid around the agent's position
+        (r, c) = agent.position
+        r_lower = max([0, r - agent.view_range])
+        r_upper = min([self.rows - 1, r + agent.view_range]) + 1
+        c_lower = max([0, c - agent.view_range])
+        c_upper = min([self.cols - 1, c + agent.view_range]) + 1
+        local_grid[
+            (r_lower+agent.view_range-r):(r_upper+agent.view_range-r),
+            (c_lower+agent.view_range-c):(c_upper+agent.view_range-c)
+        ] = self.grid[r_lower:r_upper, c_lower:c_upper]
+
+        # Generate an observation mask. The agent's observation can be blocked
+        # by other view-blocking agents, which hide the cells "behind" them. We
+        # calculate the blocking by drawing rays from the center of the agent's
+        # position to the edges of the other agents' cell. All cells that are "behind"
+        # that cell and between the two rays are invisible to the observing agent.
+        # In the mask, 1 means that the cell is visibile, 0 means that it is
+        # invisible.
+        mask = np.ones((2 * agent.view_range + 1, 2 * agent.view_range + 1))
+        for other in self.agents.values():
+            if other.view_blocking:
+                r_diff, c_diff = other.position - agent.position
+                # Ensure the other agent within the view range
+                if -agent.view_range <= r_diff <= agent.view_range and \
+                        -agent.view_range <= c_diff <= agent.view_range:
+                    if c_diff > 0 and r_diff == 0: # Other is to the right of agent
+                        upper = lambda t: (r_diff + 0.5) / (c_diff - 0.5) * t
+                        lower = lambda t: (r_diff - 0.5) / (c_diff - 0.5) * t
[…]
