### 2021-08-17 1a74611
Multi observer get_obs produces multiple grids
