### 2021-08-17 726ac48
Fixed encoding check
