### 2021-08-21 38ba48e
Adding logic for overlap matrix
