### 2021-08-21 0f3f904
Remvoed overlappable propety from agent since we use the overlap matrix now
