### 2021-08-21 30c32e1
Overlap matrix is given to grid and query logic using exceptions
