### 2021-08-21 d28e063
Updated examlpes to work with new overlapping matrix
