### 2021-08-23 #3424 @danielskatz
It appears the paper is the same in the branch and main, so I don't think it matters, but perhaps @drvinceknight or @rusu24edward could explain before we proceed?  In the meantime, I'll proofread the paper (and if changes are needed, suggest them in main)
