### 2021-08-23 #3424 @danielskatz
👋 @rusu24edward - in any case (ignoring the branch vs main issue), there are a couple of small changes needed that I've indicated in https://github.com/LLNL/Abmarl/pull/195

If the paper in main is ok, then these can just be merged and we can proceed to publication with that branch of the paper.  If the paper in the other branch is different, these changes can be made there and we can publish using the paper in that branch. But again, I really don't understand what's going on with these two branches, as the paper appears the same in both, and that's the only thing that JOSS uses the branch for.
