### 2021-08-23 #3424 @rusu24edward
> It appears the paper is the same in the branch and main, so I don't think it matters, but perhaps drvinceknight or rusu24edward could explain before we proceed? In the meantime, I'll proofread the paper (and if changes are needed, suggest them in main)

@danielskatz I'm sorry for the confusion, I'll do my best to explain. Our software utilizes a release strategy similar to the [branch release stragey](https://nvie.com/posts/a-successful-git-branching-model/). We have our main branch, which contains active development that is not yet ready for release, and when we make releases we create a release branch off main. This review has been of our first release, which is on abmarl-87-interface-release, not the main branch. The paper is the same on both branches.

Submitting the changes to main is fine, I will duplicate them for the release branch.
