### 2021-08-23 838f9ed
Minor changes to paper
