### 2021-08-23 #3424 @danielskatz
The JOSS publication is of the paper, and it includes a pointer to the full repository, so there's no need to worry about the branch in this case, I don't think. Does this make sense?
